% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ecocomDP.R
\name{aggregate_ecocomDP}
\alias{aggregate_ecocomDP}
\title{Aggregate ecocomDP data packages}
\usage{
aggregate_ecocomDP(package.ids, path = NULL)
}
\arguments{
\item{package.ids}{(character) A vector of ecocomDP data package IDs to aggregate. Use 
`view_all_ecocomDP` to get the authoritative list of available ecocomDP.
NOTE: Previous data package revisions of ecocomDP packages are valid 
entries.}

\item{path}{(character) Path to the directory in which the aggregated ecocomDP will 
be written, if specified.}
}
\value{
A single ecocomDP containing all contents of the individual ecocomDP
    listed in the "package.ids" argument.
}
\description{
Aggregate one or more ecocomDP data packages into one large ecocomDP.
}
