% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_validate.R
\name{annotate_element}
\alias{annotate_element}
\title{Annotate an EML element}
\usage{
annotate_element(element, eml, anno, rp)
}
\arguments{
\item{element}{(character) Element to be annotated}

\item{eml}{(emld list) EML to be annotated}

\item{anno}{(data.frame) The annotations.txt template}

\item{rp}{(data.frame) The responsibleParty table}
}
\value{
\item{eml}{emld list object being annotated}
}
\description{
This is a helper function for \code{annotate_eml()}.
}
\details{
Match annotatable elements (subjects) to their annotations listed in
the annotations.txt template. Target EML sub-trees where the subjects 
occur, extract their values and, when not unique, their context, then 
use these as keys to annotations.txt from which IDs and predicate + 
object annotations are retrieved and assigned to the EML. Annotatable
elements only supported through ID references are added to 
/eml/dataset/annotations (e.g. ResponsibleParty).
}
