% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_validate.R
\name{annotate_eml}
\alias{annotate_eml}
\title{Annotate an EML file (.xml) or an emld list object}
\usage{
annotate_eml(annotations = NULL, eml.in = NULL, eml.out = NULL)
}
\arguments{
\item{annotations}{(character; data frame) Path or data frame of the annotations.txt 
template for annotating \code{eml.in}. Create annotations.txt with 
\code{template_annotations()}.}

\item{eml.in}{(EML; emld list) Path to, or emld list of, the EML file to be annotated. 
Accepted versions of the emld list object can be created by
\code{EMLassemblyline::read_eml()} or \code{make_eml()}.}

\item{eml.out}{(character) Path of the annotated EML file to be written.}
}
\value{
(EML; emld list) If the input to \code{eml.in} is an EML file then the
    output will be an EML file. The emld list object is always returned.
}
\description{
Annotate an EML file (.xml) or an emld list object
}
\note{
All annotated elements are assigned ids and their annotations are 
    placed both immediately under the parent element (subject) when 
    supported and within the /eml/annotations node through id+reference 
    pairs. This redundant approach supports variation in where EML metadata 
    consumers harvest this information and supports annotation of EML 
    elements requiring id+references.
}
