% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_dictionary.R
\name{annotation_dictionary}
\alias{annotation_dictionary}
\title{Annotations of published data}
\usage{
annotation_dictionary()
}
\description{
View the collection of dataset- and attribute-level annotations
    from existing ecocomDP datasets.
}
\details{
Use the search field to find the annotation terms and URIs.
}
\note{
Access to EDI repository endpoints requires authentication. Set the environment variable \code{EDI_API_KEY} (e.g., via \code{Sys.setenv(EDI_API_KEY = "your_key")} or in your \code{.Renviron} file).

This function may not work between 01:00 - 03:00 UTC on Wednesdays due to regular maintenance of the EDI Data Repository.
}
\examples{
\dontrun{
View(annotation_dictionary())
}
    
}
