% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ants_L1}
\alias{ants_L1}
\title{The ecocomDP (L1) version of EDI data package knb-lter-hfr.118.33}
\format{
A list of:
\describe{
  \item{id}{The dataset identifier}
  \item{metadata}{Metadata for data package edi.193.5}
  \item{tables}{A list of data frames, each an ecocomDP table}
  \item{validation_issues}{Is NULL because there are no validation issues for this dataset}
}
}
\source{
Data package edi.193.5 from the Environmental Data Initiative.
}
\usage{
ants_L1
}
\description{
The the ecocomDP (L1) formatted version of EDI data package knb-lter-hfr.118.33 (Ant Assemblages in Hemlock Removal Experiment at Harvard Forest since 2003) read from the EDI API with \code{read_data(id = "edi.193.5")}. Use this dataset as an input to data "use" functions.
}
\keyword{datasets}
