% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{api_get_provenance_metadata}
\alias{api_get_provenance_metadata}
\title{Get provenance metadata}
\usage{
api_get_provenance_metadata(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") EML metadata.
}
\description{
Add Provenance Metadata from Level-1 metadata in PASTA to an XML 
    document containing a single methods element in the request message 
    body.
}
