% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{api_list_data_package_revisions}
\alias{api_list_data_package_revisions}
\title{List data package revisions}
\usage{
api_list_data_package_revisions(scope, identifier, filter = NULL, 
    environment = 'production')
}
\arguments{
\item{scope}{(character) Data package scope (e.g. 'edi', 'knb-lter-bnz').}

\item{identifier}{(character) Data package identifier (e.g. '100', '275').}

\item{filter}{(character) Filter to data packages, can be 'oldest' or 'newest'.}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Vector of revisions, if more than one exists, otherwise a 
    single revision number.
}
\description{
List Data Package Revisions operation, specifying the scope and 
    identifier values to match in the URI. The request may be filtered by 
    applying the modifiers “oldest” or “newest” to the “filter” query 
    parameter.
}
