% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{api_read_data_entity_name}
\alias{api_read_data_entity_name}
\title{Read data entity name}
\usage{
api_read_data_entity_name(package.id, identifier, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{identifier}{(character) Data entity identifier (e.g. 
5c224a0e74547b14006272064dc869b1)}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity name
}
\description{
Read Data Entity Name operation, specifying the scope, identifier, 
    revision, and entity identifier of the data entity whose name is 
    to be read in the URI.
}
