% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{api_read_data_package}
\alias{api_read_data_package}
\title{Read data package}
\usage{
api_read_data_package(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Reference URLs.
}
\description{
Read Data Package operation, specifying the scope, identifier, and 
    revision of the data package to be read in the URI, returning a resource 
    map with reference URLs to each of the metadata, data, and quality 
    report resources that comprise the data package.
}
