% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{api_read_metadata}
\alias{api_read_metadata}
\title{Read metadata}
\usage{
api_read_metadata(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') EML metadata.
}
\description{
Read Metadata (EML) operation, specifying the scope, identifier, and 
    revision of the EML document to be read in the URI.
}
