% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_validate.R
\name{assign_responsible_party_id}
\alias{assign_responsible_party_id}
\title{Update the rp object with ID information}
\usage{
assign_responsible_party_id(responsible.party, anno, rp)
}
\arguments{
\item{responsible.party}{(emld list) A ResponsibleParty EML node in the emld list format.}

\item{anno}{(data.frame) The annotations template created within 
\code{annotate_eml()}.}

\item{rp}{(data.frame) A growing set of responsible party metadata to be used
in \code{annotate_eml()}}
}
\value{
\item{responsible.party}{(emld list) Annotated responsible party node}
    \item{rp}{(data.frame) the rp object}
}
\description{
This is a helper function for \code{annotate_eml()}
}
