% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary.R
\name{calc_geo_extent_bounding_box_m2}
\alias{calc_geo_extent_bounding_box_m2}
\title{Calculate geo_extent_bounding_box_m2 for the dataset_summary table}
\usage{
calc_geo_extent_bounding_box_m2(west, east, north, south)
}
\arguments{
\item{west}{(numeric) West longitude in decimal degrees and negative if west of the prime meridian.}

\item{east}{(numeric) East longitude in decimal degrees and negative if west of the prime meridian.}

\item{north}{(numeric) North latitude in decimal degrees and negative if south of the equator.}

\item{south}{(numeric) South latitude in decimal degrees and negative if south of the equator.}
}
\value{
(numeric) Area of study site in meters squared.
}
\description{
Calculate geo_extent_bounding_box_m2 for the dataset_summary table
}
