% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary.R
\name{calc_number_of_years_sampled}
\alias{calc_number_of_years_sampled}
\title{Calculate number_of_years_sampled of the dataset_summary table}
\usage{
calc_number_of_years_sampled(dates)
}
\arguments{
\item{dates}{(Date) A vector of dates}
}
\value{
(numeric) Number of survey years
}
\description{
Calculate number_of_years_sampled of the dataset_summary table
}
\examples{

}
