% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary.R
\name{calc_std_dev_interval_betw_years}
\alias{calc_std_dev_interval_betw_years}
\title{Calculate std_dev_interval_betw_years of the dataset_summary table}
\usage{
calc_std_dev_interval_betw_years(dates)
}
\arguments{
\item{dates}{(Date) Dates from the L0 source dataset encompassing the entire study duration.}
}
\value{
(numeric) Number of survey years
}
\description{
Calculate std_dev_interval_betw_years of the dataset_summary table
}
\examples{

}
