% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data_neon.R
\name{char2datetime}
\alias{char2datetime}
\title{Parse character string to POSIXct POSIXt}
\usage{
char2datetime(datetimes)
}
\arguments{
\item{datetimes}{(character) Date or datetime}
}
\value{
(POSIXct POSIXt) \code{datetimes} parsed by relevant \code{lubridate} function
}
\description{
Parse character string to POSIXct POSIXt
}
