% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_taxon.R
\name{clean_taxon}
\alias{clean_taxon}
\title{Clean taxon (WARNING: this function is not yet functional)}
\usage{
clean_taxon(path, parent.package.id, table.name, taxon.col)

    Run this function to clean up the taxonomy data of an input dataset
    prior to running this information through the Integrated Taxonomic 
    Information System (ITIS) to obtain rank values and a Taxonomic Serial
    Number.
}
\arguments{
\item{path}{A path to the dataset working directory where the clean taxon table will
be exported.}

\item{parent.package.id}{Identifier of parent data package (e.g. knb-lter-hfr.118.28).}

\item{table.name}{Full name of table from which the taxonomy data is stored.}

\item{taxon.col}{Column names containing the taxonomy data.}
}
\value{
A tab delimited UTF-8 file in the ecocomDP working directory titled 
    \emph{taxon_clean.txt} containing clean taxon names that can then be
    used to query ITIS.
}
\description{
Clean taxonomy data of level-0 dataset and export to file.
}
\details{
This function overwrites
    \emph{taxon_clean.txt} files you have created in the ecocomDP 
    working directory.
}
