% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_workspace.R
\name{clear_workspace}
\alias{clear_workspace}
\title{Clear contents of path and close open applications and connections}
\usage{
clear_workspace(path)
}
\arguments{
\item{path}{(character) Directory in which tables, scripts, and metadata exist.}
}
\value{
(character) Message of cleaning status
}
\description{
Clear contents of path and close open applications and connections
}
\examples{

}
