% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cnvmt}
\alias{cnvmt}
\title{Convert \code{character(0)} to \code{""}}
\usage{
cnvmt(txt)
}
\arguments{
\item{txt}{(character) Character string}
}
\value{
(character) Return \code{txt} if not empty else ""
}
\description{
Convert \code{character(0)} to \code{""}
}
