% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{coerce_all_merged}
\alias{coerce_all_merged}
\title{Coerce columns of flatten_data()'s "all_merged" to correct classes}
\usage{
coerce_all_merged(x)
}
\arguments{
\item{x}{(data.frame) The all_merged object}
}
\value{
(data.frame) \code{x} with column classes coerced to ecocomDP specifications or numeric if not an ecocomDP column and if no NAs are generated.
}
\description{
Coerce columns of flatten_data()'s "all_merged" to correct classes
}
