% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_attributes.R
\name{compile_attributes}
\alias{compile_attributes}
\title{Compile attributes}
\usage{
make_eml(path, parent.package.id, child.package.id)
}
\arguments{
\item{path}{A path to the dataset working directory containing the validated 
ecocomDP tables.}

\item{delimiter}{The delimiter of input files. Can be comma (i.e. ",") or tab (i.e. "\t")}
}
\value{
Attributes formatted for make_eml.R
}
\description{
This is a helper function for make_eml.R. It compiles attributes, 
    retrieves minimum and maximum values for numeric data and reformats the 
    attributes table.
}
