% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_attributes_neon.R
\name{compile_attributes_neon}
\alias{compile_attributes_neon}
\title{compile_attributes_neon}
\usage{
make_eml_neon(x)
}
\arguments{
\item{x}{(list of data frames) The ecocomDP tables for the NEON data product.
Each item of the list must be named after the ecocomDP tables (e.g.
observation, taxon, dataset_summary, etc.).}
}
\value{
Attributes formatted for `make_eml_neon`.
}
\description{
This is a helper function for `make_eml_neon.`` It compiles attributes, 
    retrieves minimum and maximum values for numeric data and reformats the 
    attributes table.
}
