% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{create_config}
\alias{create_config}
\title{Create config.R}
\usage{
create_config(path)
}
\arguments{
\item{path}{(character) Path to which config.R will be created}
}
\value{
(logical) TRUE if written to \code{path} otherwise FALSE
}
\description{
Create config.R
}
