% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observation_ancillary.R
\name{create_observation_ancillary}
\alias{create_observation_ancillary}
\title{Create the observation_ancillary table}
\usage{
create_observation_ancillary(
  L0_wide,
  observation_id,
  variable_name,
  unit = NULL
)
}
\arguments{
\item{L0_wide}{(data.frame) The fully joined source L0 dataset, in wide format.}

\item{observation_id}{(character) Column in \code{L0_wide} containing the identifier assigned to each unique observation.}

\item{variable_name}{(character) Columns in \code{L0_wide} containing the ancillary observation data.}

\item{unit}{(character) An optional column in \code{L0_wide} containing the units of each \code{variable_name} following the column naming convention: unit_<variable_name> (e.g. "unit_temperature").}
}
\value{
(data.frame) The observation_ancillary table.
}
\description{
Create the observation_ancillary table
}
\details{
This function collects specified columns from \code{L0_wide}, converts into long (attribute-value) form with \code{variable_name} names and values to the resulting table's "variable_name" and "value" columns, respectively. Regular expression matching joins \code{unit} to any associated \code{variable_name} and is listed in the resulting table's "unit" column.
}
\examples{
wide <- ants_L0_wide

observation_ancillary <- create_observation_ancillary(
  L0_wide = wide,
  observation_id = "observation_id", 
  variable_name = c("trap.type", "trap.num", "moose.cage"))

observation_ancillary

}
