% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_to_L2_DwCA.R
\name{create_table_dwca_occurrence_core}
\alias{create_table_dwca_occurrence_core}
\title{Function creates DwC-A tables in occurrence core}
\usage{
create_table_dwca_occurrence_core(
  dwca_occurrence_core_config,
  dwca_occurrence_core_mapping,
  dt_obs,
  dt_loc,
  dt_tax,
  dt_loc_ancil = NULL,
  parent.package.id
)
}
\arguments{
\item{dwca_occurrence_core_config}{}

\item{dwca_occurrence_core_mapping}{}

\item{dt_obs}{}

\item{dt_loc}{}

\item{dt_tax}{}

\item{dt_loc_ancil}{}

\item{parent.package.id}{(character) ID of an ecocomDP data package. Only 
EDI Data Repository package IDs are currently supported.}
}
\value{

}
\description{
Function creates DwC-A tables in occurrence core
}
