% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dwca.R
\name{create_tables_dwca_event_core}
\alias{create_tables_dwca_event_core}
\title{Creates DwC-A tables in event core}
\usage{
create_tables_dwca_event_core(
  dt_obs,
  dt_obs_ancil,
  dt_loc_ancil,
  dt_loc,
  dt_tax,
  source_id,
  derived_id
)
}
\arguments{
\item{dt_obs}{(data.frame) Observation table}

\item{dt_obs_ancil}{(data.frame) Observation ancillary table}

\item{dt_loc_ancil}{(data.frame) Location ancillary table}

\item{dt_loc}{(data.frame) Location table}

\item{dt_tax}{(data.frame) Taxon table}

\item{source_id}{(character) ID of an ecocomDP data package. Only EDI Data Repository package IDs are currently supported.}

\item{derived_id}{(character) Identifier of the DwC-A dataset being created.}
}
\value{
(.csv) event, occurrence, measurementOrFact tables
}
\description{
Creates DwC-A tables in event core
}
