% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxon_ancillary.R
\name{create_taxon_ancillary}
\alias{create_taxon_ancillary}
\title{Create the taxon_ancillary table}
\usage{
create_taxon_ancillary(
  L0_wide,
  taxon_id,
  datetime = NULL,
  variable_name,
  unit = NULL,
  author = NULL
)
}
\arguments{
\item{L0_wide}{(data.frame) The fully joined source L0 dataset, in wide format.}

\item{taxon_id}{(character) Column in \code{L0_wide} containing the identifier assigned to each unique organism at the observation level.}

\item{datetime}{(character) An optional in \code{L0_wide} containing the date, and if applicable time, of ancillary location data following the ISO-8601 standard format (i.e. YYYY-MM-DD hh:mm:ss).}

\item{variable_name}{(character) Columns in \code{L0_wide} containing the ancillary taxon data.}

\item{unit}{(character) An optional column in \code{L0_wide} containing the units of each \code{variable_name} following the column naming convention: unit_<variable_name> (e.g. "unit_average_length").}

\item{author}{(character) An optional column in \code{L0_wide} containing the person associated with identification of taxa in the taxon table.}
}
\value{
(data.frame) The taxon_ancillary table.
}
\description{
Create the taxon_ancillary table
}
\details{
This function collects specified columns from \code{L0_wide}, converts into long (attribute-value) form with \code{variable_name} names and values to the resulting table's "variable_name" and "value" columns, respectively. Regular expression matching joins \code{unit} to any associated \code{variable_name} and is listed in the resulting table's "unit" column.
}
\examples{

}
