% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables}
\alias{define_variables}
\title{Define categorical variables of ecocomDP tables}
\usage{
define_variables(data.path, parent.pkg.id)
}
\arguments{
\item{data.path}{(character) Path to ecocomDP tables.}

\item{parent.pkg.id}{(character) Parent data package ID from the EDI Data Repository (e.g. 
knb-lter-cap.627.3).}
}
\value{
(data frame) A data frame with columns:
    \itemize{
        \item{tableName} Table in which variable is found.
        \item{attributeName} Field in which variable is found.
        \item{code} Variable name.
        \item{definition} Variable definition, if found, blank if otherwise.
        \item{unit} Variable unit, if found, blank otherwise.
    }
}
\description{
Get variable definitions and units from parent EML.
}
