% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{delete_from_queue}
\alias{delete_from_queue}
\title{Delete item from ecocomDP-listener queue}
\usage{
delete_from_queue(index, id)
}
\arguments{
\item{index}{(integer) Index of item to remove}

\item{id}{(character) Data package identifier, corresponding with \code{index}, to remove}
}
\value{
(logical) Indicates whether the item was successfully removed
}
\description{
Delete item from ecocomDP-listener queue
}
