% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_data_type}
\alias{detect_data_type}
\title{Detect the type of object input to the \code{data} parameter}
\usage{
detect_data_type(data)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) The dataset object returned by \code{read_data()}, or a table containing the columns of the observation table.}
}
\value{
(character) The type of \code{data}, which is one of:
\itemize{
  \item "dataset": The default return of \code{read_data()}
  \item "list_of_datasets": > 1 "dataset"
  \item "table": A flat table
  \item "list_of_tables": The named list of L1 tables (i.e. read_data()$tables)
  \item "dataset_old": The old, and since deprecated, return of \code{read_data()}
  \item "list_of_datasets_old": > 1 "dataset_old"
  \item "unrecognized": Not a recognized (i.e. supported) type
}
}
\description{
The \code{data} parameter, used by a few functions in the ecocomDP package, can accept different object types. \code{detect_data_type()} identifies the object type, which the calling function typically uses in some flow control logic.
}
