% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_delimeter.R
\name{detect_delimeter}
\alias{detect_delimeter}
\title{Detect field delimeter}
\usage{
detect_delimeter(path = "", data.files = c("data.file.1", 
    "data.file.2", "etc.") , os = "")
}
\arguments{
\item{path}{A character string specifying a path to the dataset working directory
containing the data files from which to detect the field delimeters.}

\item{data.files}{A list of character strings specifying the names of the data files
from which to detect the field delimeters of.}

\item{os}{A character string specifying the operating system in which this 
function is to be called. Valid options are generated from 
\code{detect_os}.}
}
\value{
A character string representation of the field delimeters listed in 
    order of files listed in the data.files argument.
    \item{"\\t"}{tab}
    \item{","}{comma}
    \item{";"}{semi-colon}
    \item{"|"}{pipe}
}
\description{
Detect field delimiter of input files and return character string 
    representation.
}
