% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_delimiter}
\alias{detect_delimiter}
\title{Detect field delimiter of file}
\usage{
detect_delimiter(path, data.files, os)
}
\arguments{
\item{path}{(character) Path in which \code{data.files} are found}

\item{data.files}{(character) File names}

\item{os}{(character) Return from \code{detect_os()}.}
}
\value{
(character) Field delimiter of \code{data.files}
}
\description{
Detect field delimiter of file
}
\details{
Parses the verbose return from \code{data.table::fread()} to get the delimiter value. If this fails, then a secondary function is called utilizing the suggested \code{reader} package. If this secondary approach fails, then a default "," is returned.
}
