% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{dup_fxn}
\alias{dup_fxn}
\title{Helper to handle duplicates}
\usage{
dup_fxn(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A vector of values, averaged or concatenated
}
\description{
Helper to handle duplicates
}
\details{
Helper function to handle duplicates within the observation_ancillary table. Modifications to the ecocomDP model and corresponding validation checks prohibit duplicates within ancillary tables, but this function handles instances where a user flattens data that could not otherwise be flattened.
}
