% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_template_arguments.R
\name{EAL_template_arguments}
\alias{EAL_template_arguments}
\title{Create a list of EML func inputs}
\usage{
EAL_template_arguments(
  path = NULL,
  data.path = NULL,
  data.table = NULL,
  other.entity = NULL,
  sep = NULL,
  empty = FALSE
)
}
\arguments{
\item{path}{(character) Path to the directory containing \code{make_eml()} 
metadata templates.}

\item{data.path}{(character) Path to the directory containing \code{data.table} and 
\code{other.entity} data objects.}

\item{data.table}{(character) File names of data tables. If more than one, then supply as 
a vector (e.g. \code{data.table = c('decomp.csv', 'nitrogen.csv')}).}

\item{other.entity}{(character) File names of other entity data objects. If more than one, 
then supply as a vector (e.g. 
\code{other.entity = c('ancillary_data.zip', 'processing_and_analysis.R')}).}

\item{sep}{(character) Data table field delimiter. Use this if this function fails
to read your \code{data.table}.}

\item{empty}{(logical) Initialize the output with a set of empty metadata templates?
This option is useful when wanting to transfer metadata directly into
the template files without having to call the templating functions.}
}
\value{
(named list) A list of all \code{make_eml()} arguments, specified 
    metadata templates and data objects.
}
\description{
Initialize the full list of possible inputs to \code{make_eml()}
    functions (i.e. function arguments, metadata templates, and data 
    objects). This enables bridging of upstream metadata and data sources 
    (e.g. metabases, databases, Excel files, web forms, etc.).
}
\details{
Character encoding in tabular metadata templates is converted to UTF-8 
    via \code{enc2utf8()}. Characters in TextType metadata templates are not
    yet converted. Note: This may lead to an inaccuracy and disconnect 
    between values in the data objects and what is reported in the EML (e.g.
    a categorical variable listed in the EML may not be the same as it's 
    corresponding value in the data object). To reduce the chance of this,
    warnings are issued if the input data object from which the metadata was
    extracted is not UTF-8 encoded.
}
