% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data_app.R
\name{explore_data_app}
\alias{explore_data_app}
\title{Explore data shiny app}
\usage{
explore_data_app(
  my_NEON_TOKEN = NA_character_,
  my_NEON_TOKEN_name = "NEON_TOKEN"
)
}
\arguments{
\item{my_NEON_TOKEN}{(character, default: NA_character_) NEON API token. If none is provided, the function will check if you have an API token stored as a system environmental variable set up using the directions here: https://www.neonscience.org/neon-api-tokens-tutorial}

\item{my_NEON_TOKEN_name}{(character, default: "NEON_TOKEN") Name of system environmental variable where the function will check for a NEON API token}
}
\value{
Function call runs the shiny app, does not return an R object
}
\description{
Wrapper function to run the explore data shiny app locally
}
\examples{
\dontrun{
# Run explore data shiny app
ecocomDP::explore_data_app()
}

}
