% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{Join all tables of an ecocomDP dataset and spread into wide format}
\usage{
flatten(tables)
}
\arguments{
\item{tables}{(list) A named list of ecocomDP tables (as data.frames).}
}
\value{
(data.frame) A single wide table created by joining and spreading all \code{tables}.
}
\description{
Join all tables of an ecocomDP dataset and spread into wide format
}
\examples{
\dontrun{
# Flatten NEON data
all_data <- read(
id = "DP1.20120.001",
site= c('COMO','LECO'),
startdate = "2019-06",
enddate = "2019-09",
token = Sys.getenv("NEON_TOKEN"))

my_list_of_tables <- all_data[[1]]$tables
flat_data <- flatten(my_list_of_tables)

# Flatten EDI data
all_data <- read("edi.290.1")
my_list_of_tables <- all_data[[1]]$tables
flat_data <- flatten(my_list_of_tables)
}

}
