% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_ancillary}
\alias{flatten_ancillary}
\title{Flatten ancillary table}
\usage{
flatten_ancillary(ancillary_table)
}
\arguments{
\item{ancillary_table}{(data.frame) Ancillary table to flatten. Can be: observation_ancillary, location_ancillary, or taxon_ancillary.}
}
\value{
(data.frame) A flattened version of \code{ancillary_table} with units added following the unit_<variable_name> convention, and sorted so the unit immediately follows its corresponding variable_name.
}
\description{
Flatten ancillary table
}
