% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data}
\alias{flatten_data}
\title{Join all tables of an ecocomDP dataset and spread into flat format}
\usage{
flatten_data(tables)
}
\arguments{
\item{tables}{(list) A named list of ecocomDP tables as data.frames}
}
\value{
(data.frame) A single flat table created by joining and spreading all \code{tables} except observation
}
\description{
Join all tables of an ecocomDP dataset and spread into flat format
}
\examples{
dataset <- ants_L1

flat <- flatten_data(dataset[[1]]$tables)

flat

}
