% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flatten_location}
\alias{flatten_location}
\title{Unnest nested sites in the location table}
\usage{
flatten_location(dt_loc)
}
\arguments{
\item{dt_loc}{(data.frame) The location table}
}
\value{
(data.frame) With columns location_id, site_name, latitude, longitude, 
    elevation.
}
\description{
Unnest nested sites in the location table
}
\details{
Dereferences each of the parent_id, creates a site name that includes
    the site names of each parent, and adds the lowest latitude, longitude 
    and elevation available in the table.
}
