% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{format_for_comm_plots}
\alias{format_for_comm_plots}
\title{Format dataset for community plotting functions}
\usage{
format_for_comm_plots(observation, id)
}
\arguments{
\item{observation}{(data.frame) The observation table of ecocomDP}

\item{id}{(character) Dataset id. Used in plot titles.}
}
\value{
(data.frame) Tabular data of \code{id} in a format compatible with plotting functions
}
\description{
Format dataset for community plotting functions
}
\details{
Downsteam plotting functions are based on \href{https://github.com/sokole/ltermetacommunities/tree/master/Group2-explore-data}{LTER Metacommunities code} and use their intermediate data input format.
}
