% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_popler_2_observation.R
\name{format_popler_2_observation}
\alias{format_popler_2_observation}
\title{Reformat popler data to ecocomDP observation table}
\usage{
format_popler_2_observation(path.in = "", data.list = c("data.list.1", 
    "data.list.2", "etc."), path.out = "")
}
\arguments{
\item{path.out}{A character string specifying a path to the directory to which the 
reformatted popler datasets will be exported to.}

\item{data.list}{A list of character strings specifying the names of the data list files.}
}
\value{
Data tables (.csv) in the ecocomDP format to the directory specified
    by the path.out argument.
}
\description{
Load datasets from the popler database 
    (https://github.com/AldoCompagnoni/popler) and reformat to the 
    observation table of the ecocomDP project
    (https://github.com/EDIorg/ecocomDP/tree/master).
}
