% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{format_search_results}
\alias{format_search_results}
\title{Reformat search results}
\usage{
format_search_results(x)
}
\arguments{
\item{x}{(data.frame) Current output of \code{search_data()}}
}
\value{
(data.frame) New output of \code{search_data()} reformatted to support L1 versioning
}
\description{
This temporarily accomodates NEON's plan for multiple L1 versions and will have to be refactored to support other data sources adopting this use case. A permanent solution is to refactor the \code{summarize_data_neon()} and \code{summarize_data_edi()} functions to return the exact objects of interest.
}
\details{
Joins /inst/L1_versions.txt (a file holding version info) with \code{x} and does a little reformatting.
}
