% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_attr_defs}
\alias{get_attr_defs}
\title{Get attribute definitions from EML}
\usage{
get_attr_defs(eml)
}
\arguments{
\item{eml}{(xml_document, xml_node) EML metadata}
}
\value{
(named list) Definitions
}
\description{
Get attribute definitions from EML
}
\note{
Duplicate names are dropped.
}
