% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{get_authority}
\alias{get_authority}
\title{Get taxonomic authority}
\usage{
get_authority(taxon, data.source)
}
\arguments{
\item{taxon}{A character string representation of the taxon to search on.}

\item{data.source}{A numeric ID corresponding to the data source (i.e. taxonomic authority)
you'd like to query. Run `view_authorities` to get valid data source
options and ID's.}
}
\value{
\itemize{
        \item{resolved_name} Resolved taxon name.
        \item{authority} Name of the authority searched.
        \item{score} Relative match score provided by the authority.
    }
}
\description{
Use fuzzy searching in the Global Names Resolver to correct spelling
    and locate appropriate authorities.
}
