% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{get_classification}
\alias{get_classification}
\title{Get the taxonomic classification hierarchy for taxa resolved to supported authorities}
\usage{
get_classification(
  taxa.clean,
  authority = NA,
  authority.id = NA,
  rank = NA,
  path = NULL
)
}
\arguments{
\item{taxa.clean}{(character) Taxa names}

\item{authority}{(character) Authority \code{taxa.clean} have been resolved to, otherwise \code{NA}. Supported authorities include: "ITIS", "WORMS", "GBIF".}

\item{authority.id}{(character) ID of \code{taxa.clean} within the \code{authority}, otherwise \code{NA}}

\item{rank}{(character) Rank (e.g. "Genus", "Species") of \code{taxa.clean}, otherwise \code{NA}. This is useful when \code{taxa.clean} can't be resolved to an \code{authority} and the rank must be manually defined.}

\item{path}{(character) Path of the directory containing taxa_map.csv.}
}
\value{
(list) For \code{taxa.clean} resolved to a supported authority, each item in the list is a classification hierarchy (also a list), including one or more common names (only when \code{authority} is ITIS or WORMS) and authority IDs for each rank-value pair. For \code{taxa.clean} not resolved to a supported authority, each item is listed as defined in the \code{taxa.clean}, \code{authority}, and \code{authority.id} arguments.
}
\description{
Get the taxonomic classification hierarchy for taxa resolved to supported authorities
}
\details{
Only taxa resolved to supported authorities can be expanded into a full taxonomic classification with common names. Taxa resolved to unsupported authorities, or not resolved at all, will be listed as is defined in the \code{taxa.clean}, \code{authority}, and \code{authority.id} arguments.

    Supported authorities are recognized by a controlled set of representations.
    \itemize{
    \item{ITIS can be: "ITIS", "itis", "Integrated Taxonomic Information System", or "https://www.itis.gov/".}
    \item{WORMS can be: "WORMS", "worms", "World Register of Marine Species", or "http://www.marinespecies.org/".}
    \item{GBIF can be: "GBIF", "gbif", "GBIF Backbone Taxonomy", or "https://gbif.org".}
    }
}
