% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_eml_attribute}
\alias{get_eml_attribute}
\title{Get metadata for an attribute}
\usage{
get_eml_attribute(attr.name, package.id)
}
\arguments{
\item{attr.name}{(character) Attribute name}

\item{package.id}{(character) Data package identifier (e.g. knb-lter-cap.627.3) in the 
EDI Data Repository.}

\item{evironment}{(character) Repository environment in which the L0 and L1 exist. Some repositories have development, staging, and production environments which are distinct from one another. This argument allows execution for the \code{update_L1} workflow within the context of one of these environments. Default is "production".}
}
\value{
(list) A named vector with these attribute elements:
    \itemize{
        \item{name} <attributeName>
        \item{definition} <attributeDefinition> 
        \item{unit} <standardUnit> or <customUnit>
    }
}
\description{
Get attribute metadata
}
