% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_eol}
\alias{get_eol}
\title{Get end of line (EOL) character}
\usage{
get_eol(path, file.name, os)
}
\arguments{
\item{path}{(character) A path to the target file directory.}

\item{file.name}{(character) The target file name.}

\item{os}{(character) The operating system in which this function is called
called. Valid options are generated from \code{detect_os}.}
}
\value{
A character string representation of the EOL character.
}
\description{
Get EOL character of input file(s).
}
