% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{get_from_queue}
\alias{get_from_queue}
\title{Get next item in ecocomDP-listener queue}
\usage{
get_from_queue()
}
\value{
\item{index}{(integer) Index of item in queue. Is later used for removing the item from the queue.}
\item{id}{(character) Data package identifier}
}
\description{
Get next item in ecocomDP-listener queue
}
\details{
The queue is /webapp/ecocomDP.sqlite in \href{https://github.com/EDIorg/ecocomDP-listener}{ecocomDP-listener} and is accessible via HTTP GET.
}
