% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{get_id}
\alias{get_id}
\title{Get taxonomic identifiers}
\usage{
get_id(taxon, authority)
}
\arguments{
\item{taxon}{A character string specifying taxon to get the ID for.}

\item{authority}{A character string specifying the authority from which to get the ID.}
}
\value{
\itemize{
        \item{taxon_id} An authority ID for the taxon.
        \item{rank} The taxonomic rank of the taxon.
    }
}
\description{
Get a taxonomic identifier for a taxon name and corresponding authority.
}
