% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_metadata.R
\name{get_neon_metadata}
\alias{get_neon_metadata}
\title{Read in NEON EML for a specified data product}
\usage{
get_neon_metadata(dpid)
}
\arguments{
\item{dpid}{(character) NEON data product ID (e.g. DP1.20120.001).}
}
\value{
EML metadata for a NEON data product.
}
\description{
Get metadata from the NEON data portal for a data product ID. NOTE: The
    retrieved metadata file is specific to a random site and date range. 
    This function is intended to be used with `get_neon_data` and 
    `make_neon_eml` to craft the EML for a NEON data product converted to 
    the ecocomDP.
}
