% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_data_package_id.R
\name{get_previous_data_package_id}
\alias{get_previous_data_package_id}
\title{Get previous data package identifier}
\usage{
get_previous_data_package_id(
  package.id,
  repository = "EDI",
  environment = "production"
)
}
\arguments{
\item{package.id}{(character) Data package identifier}

\item{repository}{(character) Data repository in which \code{package.id} resides and associated with \code{environment}. Currently supported repositories are: "EDI" (Environmental Data Initiative). Requests for support of other repositories can be made via \href{https://github.com/EDIorg/ecocomDP}{ecocomDP GitHub} issues. Default is "EDI".}

\item{evironment}{(character) Repository environment in which \code{package.id} exists. Some repositories have development, staging, and production environments which are distinct from one another. This argument allows reading of EML from different environments. Default is "production".}
}
\value{

}
\description{
A wrapper function to repository specific read methods (the repository arg drives the logic).
}
