% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_L0_id.R
\name{get_previous_L0_id_and_L1_conversion_script}
\alias{get_previous_L0_id_and_L1_conversion_script}
\title{Get previous L0 data package identifier and URL link to conversion script}
\usage{
get_previous_L0_id_and_L1_conversion_script(package.id)
}
\arguments{
\item{package.id}{(character) Newest L0 data package identifier}
}
\value{
(list) A list containing:
\item{id.L0}{Identifier of previous L0 data package}
\item{id.L1}{Identifier of L1 that is a descendant of L0 and from which the \code{url} was obtained}
\item{url}{Download URL of the script used to convert the L0 to L1}
Otherwise NULL
}
\description{
Get the most recent L0 data package identifier containing a recognizable L0 to L1 conversion script and get the URL to download the script as well.
}
\details{
For the most recent L0 data package, which may not have a corresponding L1, follow all provenance traces to data packages containing the "ecocomDP" keyword. If more than one, select the most recent L0 identifier, read the corresponding EML, look for a valid L0 to L1 conversion script (convert_ecocomDP.R), and break if found. Otherwise try this process again for the next oldest L0 data package and so on.
}
