% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_previous_version}
\alias{get_previous_version}
\title{Get previous data package version}
\usage{
get_previous_version(package.id)
}
\arguments{
\item{package.id}{(character) Data package identifier}
}
\value{
(character) Previous data package version
}
\description{
Get previous data package version
}
\details{
Supports repository specific methods.
}
