% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_child.R
\name{get_report_datetime}
\alias{get_report_datetime}
\title{Get datetime of data package quality report from the EDI Data Repository}
\usage{
get_report_datetime(package.id, environment)
}
\arguments{
\item{package.id}{(character) Data package identifier}

\item{environment}{(character) Data repository environment}
}
\value{
(character) Datetime in the form YYYY-MM-DDThh:mm:ss
}
\description{
Get datetime of data package quality report from the EDI Data Repository
}
