% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_child.R
\name{has_child}
\alias{has_child}
\title{Does a data package have a ecocomDP or DwC-A child?}
\usage{
has_child(keyword, package.id)
}
\arguments{
\item{keyword}{(character) Specifies the child type. Can be: "ecocomDP" or "Darwin Core Archive (DwC-A) Event Core"}

\item{package.id}{(character) Data package identifier}
}
\value{
(named logical) TRUE/FALSE named with the child's data package identifier
}
\description{
Does a data package have a ecocomDP or DwC-A child?
}
