% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_child_dwcae.R
\name{has_child_dwcae}
\alias{has_child_dwcae}
\title{Does a data package have a Darwin Core Archive Event Core child?}
\usage{
has_child_dwcae(package.id)
}
\arguments{
\item{package.id}{(character) Data package identifier}
}
\value{
(logical) TRUE/FALSE
}
\description{
Does a data package have a Darwin Core Archive Event Core child?
}
