% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_child.R
\name{has_unprocessed_versions}
\alias{has_unprocessed_versions}
\title{Check if earlier unprocessed versions of a data package are in the ecocomDP-listener queue}
\usage{
has_unprocessed_versions(package.id)
}
\arguments{
\item{package.id}{}
}
\value{
(logical) TRUE if earlier versions of a data package are found
}
\description{
The presence of such items may indicate the integrity of a data package series is compromised and processing should be halted until the issue is fixed.
}
