% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable_name.R
\name{hasvars}
\alias{hasvars}
\title{Identify tables with variable names}
\usage{
hasvars(tables, vname)
}
\arguments{
\item{tables}{(list) Data object returned by \code{read_data()[[1]]$tables}}

\item{vnames}{(character) Variable name}
}
\value{
(logical) TRUE for \code{tables} containing \code{vars}
}
\description{
Identify tables with variable names
}
\note{
Does not return TRUE for the variable_mapping table.
}
