% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{increment_package_version}
\alias{increment_package_version}
\title{Increment data package version number}
\usage{
increment_package_version(package.id)
}
\arguments{
\item{package.id}{(character) Data package identifier}
}
\value{
(character) Package identifier with version number incremented by 1.
}
\description{
Increment data package version number
}
\details{
Supports repository specific methods.
}
\examples{
increment_package_version("edi.100.1")

}
