% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_missing_libraries.R
\name{install_missing_libraries}
\alias{install_missing_libraries}
\title{Install missing libraries}
\usage{
install_missing_libraries(conversion.script)
}
\arguments{
\item{conversion.script}{(character) Full path to create_ecocomDP.R}
}
\description{
Identify if a create_ecocomDP.R script uses uninstalled libraries and install them for use in \code{update_L1()}.
}
\details{
Libraries used by create_ecocomDP.R are identified by reading the full file to memory and using regular expressions to extract the name of the library from within \code{library()} (i.e. \code{"(?<=library\\().*(?=\\))"}). All libraries found in this way are installed.
}
