% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{is_dataset}
\alias{is_dataset}
\title{Validate dataset structure}
\usage{
is_dataset(dataset)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}
}
\value{
(logical) TRUE if has dataset structure otherwise FALSE
}
\description{
Validate dataset structure
}
