% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_edi}
\alias{is_edi}
\title{Does dataset ID belong to EDI?}
\usage{
is_edi(id)
}
\arguments{
\item{id}{(character) Dataset identifier}
}
\value{
(logical) TRUE if is an EDI identifier, otherwise FALSE
}
\description{
Does dataset ID belong to EDI?
}
