% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_empty_nodeset}
\alias{is_empty_nodeset}
\title{Is empty nodeset?}
\usage{
is_empty_nodeset(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Any nodeset returned by the xml2 library}
}
\value{
(logical) TRUE if nodeset length = 0
}
\description{
Is empty nodeset?
}
