% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_neon}
\alias{is_neon}
\title{Does dataset ID belong to NEON?}
\usage{
is_neon(id)
}
\arguments{
\item{id}{(character) Dataset identifier}
}
\value{
(logical) TRUE if is a NEON identifier, otherwise FALSE
}
\description{
Does dataset ID belong to NEON?
}
