% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_prov}
\alias{is_prov}
\title{Is provenance node?}
\usage{
is_prov(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) methods nodeset at \code{/eml:eml/dataset/methods/methodStep}}
}
\value{
(logical) TRUE if nodeset has provenance
}
\description{
Is provenance node?
}
\details{
Looks for provenance in \code{./dataSource}
}
