% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{join_obs_loc_tax}
\alias{join_obs_loc_tax}
\title{Join observation, location, and taxon tables in a wide format}
\usage{
join_obs_loc_tax(dt_obs, dt_loc, dt_tax)
}
\arguments{
\item{dt_obs}{(data.frame) The observation table}

\item{dt_loc}{(data.frame) The location table}

\item{dt_tax}{(data.frame) The taxon table}
}
\value{
(data.frame) Joined observation, location, and taxon tables in a wide 
    format, where the location table has been unnested for latitude, 
    longitude, and site names.
}
\description{
Join observation, location, and taxon tables in a wide format
}
