% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_codes_to_definitions.R
\name{L0_codes_to_definitions}
\alias{L0_codes_to_definitions}
\title{Convert L0 codes to definitions}
\usage{
L0_codes_to_definitions(x, col.name, x.object.name, eml)
}
\arguments{
\item{x}{(data frame) L0 table containing the column to convert.}

\item{col.name}{(character) Name of column to convert.}

\item{x.object.name}{(character) Object name in EML of the table to convert (i.e. file name).}

\item{eml}{(xml_document, xml_node) EML metadata.}
}
\value{
(character) Vector of definitions corresponding to codes.
}
\description{
A function to convert categorical codes to definitions. Occasionally 
    L1 content is locked up in L0 code definitions (e.g. species names). 
    This function converts these codes to the associated definitions.
}
