% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_missing_value_codes_to_na.R
\name{L0_missing_value_codes_to_na}
\alias{L0_missing_value_codes_to_na}
\title{Convert L0 missing value codes to NA}
\usage{
L0_missing_value_codes_to_na(x, x.object.name, eml)
}
\arguments{
\item{x}{(data frame) L0 table with missing value codes to convert.}

\item{x.object.name}{(character) Object name in EML of the table to convert (i.e. file name).}

\item{eml}{(xml_document, xml_node) EML metadata. Use 
`EDIutils::api_read_metadata()` or `xml2::read_xml()` to read in the 
EML.}
}
\value{
(data frame) `x` with missing value codes converted to NA.
}
\description{
A function to convert missing value codes to NA. Converting missing 
    value codes to a single type simplifies downstream usage.
}
