% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_trimws.R
\name{L0_trimws}
\alias{L0_trimws}
\title{Trim white space from L0 character columns}
\usage{
L0_trimws(x)
}
\arguments{
\item{x}{(data frame) L0 table to trim white space from.}
}
\value{
(data frame) L0 table with white space removed from character class 
    columns.
}
\description{
A function to remove white space from L0 character class columns. The 
    presence of white space affects alignment with L0 EML and downstream usage.
}
