% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_to_L2_DwCA.R
\name{L1_to_L2_DwCA}
\alias{L1_to_L2_DwCA}
\title{Creates Darwin Core Archive package from an ecocomDP formatted level-1}
\usage{
L1_to_L2_DwCA(core.name, L1.id)
}
\arguments{
\item{core.name}{(character) The Darwin Core central table of the package. Can be: 
"occurence" (occurrence core) or "event" (event core).}

\item{L1.id}{(character) Package ID of an ecocomDP dataset from PASTA. Other level-1 
types TBD.}
}
\value{
Data entities in either the occurrence or event core formats with a
    corresponding EML metadata record and a meta.xml file.
}
\description{
Creates Darwin Core Archive package from an ecocomDP formatted level-1
}
\details{
TODO: Add details
}
\examples{

}
