% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{load_gnr_datasources}
\alias{load_gnr_datasources}
\title{Load and fix GNR Datasources}
\usage{
load_gnr_datasources()
}
\value{
(data.frame) GNR datasources from \code{taxize::gnr_datasources()}
}
\description{
Load and fix GNR Datasources
}
\details{
This fixes bugs in taxize which otherwise produce inconsistent datasource names (e.g. "Integrated Taxonomic Information SystemITIS" rather than expected "ITIS")
}
