% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_tables.R
\name{long2wide_obs_loc_tax}
\alias{long2wide_obs_loc_tax}
\title{Join observation, location, and taxon tables in a wide format}
\usage{
long2wide_obs_loc_tax(dt_obs, dt_loc, dt_tax)
}
\arguments{
\item{dt_obs}{(data.frame) The observation table}

\item{dt_loc}{(data.frame) The location table}

\item{dt_tax}{(data.frame) The taxon table}
}
\value{
(data.frame) Joined observation, location, and taxon tables in a wide 
    format, where the location table has been unnested for latitude, 
    longitude, and site names.
}
\description{
Join observation, location, and taxon tables in a wide format
}
\details{
TODO: "Denormalizd" + describe the unnesting logic.
    TODO: Intended to be used within a function that has already read the 
    data tables.
}
\examples{
\dontrun{
long2wide_obs_loc_tax(
  dt_obs = my_package_observation,
  dt_loc = my_package_location,
  dt_tax = my_package_taxon)
}

}
