% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_neon_data_availability.r
\name{lookup_neon_data_availability}
\alias{lookup_neon_data_availability}
\title{Lookup NEON data product availability}
\usage{
lookup_neon_data_availability(
  dpid = "DP1.10086.001",
  sample_location_list = NA,
  sample_location_type = "siteID",
  sample_date_min = "2012-01-01",
  sample_date_max = Sys.Date(),
  sample_date_format = "\%Y-\%m-\%d",
  data_package_type = "basic",
  url_prefix_data = "http://data.neonscience.org:80/api/v0/data/",
  url_prefix_products = "http://data.neonscience.org:80/api/v0/products/",
  ...
)
}
\arguments{
\item{dpid}{character sting for NEON data product ID}

\item{sample_location_list}{list of sites, domains, etc. If NA, retrieve all data for the given data table / dpid combination.}

\item{sample_location_type}{character sting for location type, such as 'siteID'. Must be one of the NEON controlled terms. If you're unsure, use 'siteID'}

\item{sample_date_min}{start date for query. Default is 1-Jan-2012, and this should capture the earliest NEON data record.}

\item{sample_date_max}{end date for query. Default is current date.}

\item{sample_date_format}{date format. Default/expected format is yyyy-mm-dd}

\item{data_package_type}{package type, either 'basic' or 'expanded'. If unsure, use 'expanded'}

\item{url_prefix_data}{data endpoint for NEON API.}

\item{url_prefix_products}{products endpoint for NEON API.}
}
\value{
data frame with REST calls for sites and dates that have available data
}
\description{
This is a function to look up the sites and dates 
for which data are available on the NEON data portal for a given 
data product identifier (dpid).
}
\examples{

# look up by dpid
lookup_neon_data_availability("DP1.10086.001")

# look up by dpid for a specific site (CPER)
lookup_neon_data_availability(dpid = "DP1.10086.001",
                                     sample_location_list = 'CPER')
                                     
# look up by dpid for multiple sites                                      
lookup_neon_data_availability(dpid = "DP1.10086.001",
                                     sample_location_list = c('CPER','BART'))

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Eric R. Sokol \email{esokol@battelleecology.org}
}
