% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset_summary.R
\name{make_dataset_summary}
\alias{make_dataset_summary}
\title{Make dataset_summary table}
\usage{
make_dataset_summary(
  parent.package.id,
  child.package.id,
  sample.dates,
  taxon.table
)
}
\arguments{
\item{parent.package.id}{(character) Parent data package identifier composed of scope, 
identifier, and revision (e.g. 'edi.101.1').}

\item{child.package.id}{(character) Child data package identifier composed of scope, 
identifier, and revision (e.g. 'edi.101.1').}

\item{sample.dates}{(character) Sample dates from which to calculate date fields of the
dataset_summary table.}

\item{taxon.table}{(data frame) The ecocomDP taxon table.}
}
\value{
A data frame of the dataset_summary table.
}
\description{
Make the ecocomDP dataset_summary table.
}
