% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{make_eml}
\alias{make_eml}
\title{Make EML}
\usage{
make_eml(path, import.eml = FALSE)
}
\arguments{
\item{path}{A path to the dataset working directory containing the completed 
metadata templates, \emph{eml_configuration.R}, 
\emph{datasetname_datatablename_catvars.txt} (if categorical variables 
are present), and \emph{geographic_coverage.txt} (if reporting detailed 
geographic coverage).}

\item{import.eml}{If TRUE then EML for the raw dataset is read in from the working 
directory and top level elements are used and data table specifc 
elements are replaced with those for the ecocomDP tables.}
}
\value{
Validation results printed to the \emph{Console}.
    
    An EML metadata file written to the dataset working directory titled 
    \emph{packageID.xml}.
}
\description{
Translate user supplied metadata into the EML schema for the Ecological 
    Community Data Pattern (ecocomDP), validate the schema, and write to 
    file.
}
\details{
If validation fails, open the EML document in a .xml editor to identify 
    the source of error. Often the error is small and quickly resolved with 
    the aid of an editors schema congruence checker.
}
