% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml_neon.R
\name{make_eml_neon}
\alias{make_eml_neon}
\title{Make EML for a NEON data product}
\usage{
make_eml_neon(eml, x, protocols, dp.id)
}
\arguments{
\item{eml}{(eml) EML ofa NEON data product.}

\item{x}{(list of data frames) The ecocomDP tables for the NEON data product.
Each item of the list must be named after the ecocomDP tables (e.g.
observation, taxon, dataset_summary, etc.).}

\item{protocols}{(character) Vector of character strings listing the NEON sampling 
protocols.}

\item{dp.id}{(character) NEON data product ID.}

\item{eml.path}{(character) Path to where the EML will be written.}
}
\value{
EML metadata for a NEON data product.
}
\description{
Make EML metadata for a NEON data product reformatted in the ecocomDP.
}
