% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location.R
\name{make_location}
\alias{make_location}
\title{Make the ecocomDP location table}
\usage{
make_location(x, cols, kml = NULL, col = NULL, parent.package.id = NULL)
}
\arguments{
\item{x}{(data frame) A data frame containing sampling location names (in 
columns) for each observation (rows).}

\item{cols}{(character) Columns in x containing sampling location data. The list 
order should reflect the hierarchical arrangement of sampling locations.
E.g.: If 'site' is the coarsest level of observation, 'habitat' is 
nested within 'site', 'transect' is nested within 'habitat', and 
'quadrat' is nested within 'transect' then the 'cols' argument should
be supplied as: cols = c('site', 'habitat','transect', 'quadrat').}

\item{kml}{A data object (simple feature collection) created with the `st_read``
function of the `sf`` (simple features) R package. This data object
contains a field called 'Name'. Values listed in this field must match
the corresponding values listed in the data frame (x) and specified in
'cols'.}

\item{col}{(character) The column name specified in the input argument 'cols' to 
which the data in the kml object corresponds to. E.g., if the kml 
corresponds with with the column named 'site' then enter 'site' for the 
col argument.}

\item{parent.package.id}{(character) Parent data package ID (e.g. "knb-lter-nin.9.1") from which 
geographicCoverage will be extracted. Currently only works for EDI
data.}
}
\value{
A list containing:
\item{location}{A data frame of the location table}
\item{x}{The input data frame \code{x} but with an added location_id column.
This simplifies alignment of values between \code{x} and location.}
}
\description{
Make the ecocomDP location table.
}
\note{
The arguments 'kml' and 'col' only accept single inputs. I.e. does 
    not support kml data for more than one col.
}
