% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_ancillary.R
\name{make_location_ancillary}
\alias{make_location_ancillary}
\title{Make the location ancillary table}
\usage{
make_location_ancillary(x = NULL, cols = NULL, eml = NULL)
}
\arguments{
\item{x}{(data frame) The master data frame composed of all joined tables from the 
parent data package for which the core ecocomDP tables have been, 
created. This master data frame is linked to core tables via primary 
keys and has all ecocomDP column names added to this point. This funciton 
will not work otherwise.}

\item{cols}{(character) Names of all columns from \code{x} needed to create 
location_ancillary. This function assumes that all columns with names
corresponding to the ecocomDP location_ancillary table will be treated
as such. If any \code{cols} have a different meaning, i.e. within the 
context of the parent data package, then this function will not work.}

\item{eml}{(xml_document xml_node) EML metadata listing units for variables listed 
in \code{cols}. Use \code{EDIutils::api_read_metadata()} or 
\code{xml2::read_xml()} to read the EML file.}
}
\value{
(data frame) The location_ancillary table.
}
\description{
Make the location ancillary table
}
\details{
This is possible because at this point we can distinguish between 
    standard column names and non-standard column names (variables).
    
    This function will not work if variables listed in \code{cols} have 
    duplicate listing in \code{eml}.
}
