% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxon_ancillary.R
\name{make_taxon_ancillary}
\alias{make_taxon_ancillary}
\title{Make the taxon ancillary table}
\usage{
make_taxon_ancillary(x = NULL, cols = NULL, eml = NULL)
}
\arguments{
\item{x}{(data frame) The master data frame composed of all joined tables from the 
parent data package for which the core ecocomDP tables have been, 
created. This master data frame is linked to core tables via primary 
keys and has all ecocomDP column names added at this point, specifically
relevant to the location_ancillary table are the columns "taxon_id", 
and "author" ("datetime" will be automatically created from the 
observation table "observation_datetime" column). This funciton 
will not work otherwise.}

\item{cols}{(character) Names of all columns from \code{x} needed to create 
taxon_ancillary. This function assumes that all columns with names
corresponding to the ecocomDP location_ancillary table will be treated
as such. If any \code{cols} are shared between the parent tables (now
joined in the master data frame) and taxon_antillary, then there is
ambiguous meaning and this function will not work.}

\item{eml}{(xml_document xml_node) EML metadata listing units for variables listed 
in \code{cols}. Use \code{EDIutils::api_read_metadata()} or 
\code{xml2::read_xml()} to read the EML file.}
}
\value{
(data frame) The taxon_ancillary table.
}
\description{
Make the taxon ancillary table
}
\details{
This is possible because at this point we can distinguish between 
    standard column names and non-standard column names (variables).
    
    This function will fail to resolve units for a variable if the variable 
    listed in \code{cols} has a duplicate listing in \code{eml}. In such
    cases a warning is returned with the suggestion to remove duplicate 
    variables from the EML when running this function.
}
