% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_neon_data_to_ecocomDP.ALGAE.R,
%   R/map_neon_data_to_ecocomDP.r
\name{map_neon_data_to_ecocomDP.ALGAE}
\alias{map_neon_data_to_ecocomDP.ALGAE}
\alias{map_neon_data_to_ecocomDP}
\title{Map NEON data products to ecocomDP}
\usage{
map_neon_data_to_ecocomDP.ALGAE(neon.data.product.id = "DP1.20166.001", ...)

map_neon_data_to_ecocomDP(neon.data.product.id, ...)
}
\arguments{
\item{neon.data.product.id}{The identifier of the NEON data product to pull, in the form DPL.PRNUM.REV, e.g. DP1.20166.001 (neon.data.product.id for Periphyton, seston, and phytoplankton collection)}

\item{...}{Additional arguments passed to neonUtilities::loadByProduct().}

\item{site}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all.}

\item{startdate}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to NA.}

\item{enddate}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to NA.}

\item{check.size}{T or F, should the user approve the total file size before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use check.size=F.}

\item{nCores}{The number of cores to parallelize the stacking procedure. By default it is set to a single core.}

\item{forceParallel}{If the data volume to be processed does not meet minimum requirements to run in parallel, this overrides. Set to FALSE as default.}
}
\value{
Returns a named list, including: 
\item{neon_metadata}{A list of information about the NEON data product returned by neonUtilities::getProductInfo.}
\item{ecocomDP_tables}{A list of data.frames following the ecocomDP format.}
}
\description{
Pull files from the NEON API by data product, merge data for each table, and convert to ecocomDP format. Please see neonUtilities::loadByProduct for more information.
}
\details{
All available data meeting the query criteria will be downloaded. Most data products are collected at only a subset of sites, and dates of collection vary. Consult the NEON data portal for sampling details.
Dates are specified only to the month because NEON data are provided in monthly packages. Any month included in the search criteria will be included in the download. Start and end date are inclusive.
}
\section{Functions}{
\itemize{
\item \code{map_neon_data_to_ecocomDP.ALGAE}: This method will retrieve density data for ALGAE from neon.data.product.id DP1.20166.001 (Periphyton, seston, and phytoplankton collection) from the NEON data portal and map to the ecocomDP
}}

\examples{
\dontrun{
# Retrieve and map NEON ALGAE data to ecocomDP format
my_result <- ecocomDP::map_neon_data_to_ecocomDP(
  neon.data.product.id = "DP1.20166.001",
  site = c("MAYF", "PRIN"),
  startdate = "2016-1",
  enddate = "2018-11",
  check.size = FALSE)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Eric R Sokol \email{esokol@battelleecology.org}
}
