% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_maintainers.R
\name{message_maintainers}
\alias{message_maintainers}
\title{Message maintainers}
\usage{
message_maintainers()
}
\description{
Send email to maintainers with a log file specific to the data object conversion (not the entire log_master.txt).
}
