% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{my_plot_sites3}
\alias{my_plot_sites3}
\title{Work-in-progress upgraded plot_sites}
\usage{
my_plot_sites3(
  data,
  id = NA_character_,
  alpha = 1,
  color_var = "package_id",
  shape_var = "package_id",
  size_var = 3
)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) The dataset object returned by \code{read_data()}, a named list of tables containing the observation and taxon tables, or a flat table containing columns of the observation and location tables.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles. Is automatically assigned when \code{data} is a dataset object containing the \code{id} field, or is a table containing the package_id column.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}

\item{color_var}{(character) Argument to assign colors to the points on the plot}

\item{shape_var}{(character) Argument to assign shapes to the points on the plot}

\item{size_var}{(character) Argument to assign size to the points on the plot}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Work-in-progress upgraded plot_sites
}
\details{
The \code{data} parameter accepts a range of input types but ultimately requires the 14 columns of the combined observation and location tables.
}
\examples{
\dontrun{
# Read a dataset of interest
dataset <- read_data("edi.193.5")

# Plot the dataset
plot_sites(dataset)

# Flatten dataset then plot
dataset \%>\%
  flatten_data() \%>\%
  plot_sites()
}

# Plot the example dataset
plot_sites(ants_L1)
}
