% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{not_all_NAs}
\alias{not_all_NAs}
\title{Determine if a col is all NAs}
\usage{
not_all_NAs(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
(logical) TRUE if not all NAs, otherwise FALSE
}
\description{
Determine if a col is all NAs
}
