% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot alpha diversity (species richness) over time and space}
\usage{
plot_alpha_diversity(dslong, id, path, vunit, alpha)
}
\arguments{
\item{dslong}{(data.frame) Long form of return from \code{format_for_comm_plots()}}

\item{id}{(character) Dataset id}

\item{path}{(character) Path to where plots will be written}

\item{vunit}{(character) Unit of measurement variable}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\value{
\item{alpha_diversity.pdf}{Alpha diversity (species richness) over time and space}
}
\description{
Plot alpha diversity (species richness) over time and space
}
