% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot community metrics}
\usage{
plot_data(dataset, type, path, alpha = 1)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}

\item{type}{(character) Type of plots to create. Options: "community"}

\item{path}{(character) Path to directory where plots will be written}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\value{
Returns a set of figures:
\item{alpha_diversity.pdf}{Alpha diversity (species richness) over time and space. number of unique taxa at site through time with total for all sites}
\item{sampling_effort.pdf}{Spatiotemporal sampling effort. Notes when sites were sampled.}
\item{sp_accumulation_over_space.pdf}{Species accumulation curve over space. Cummulative number of unique taxa across all dates at sites, ordered by site id}
\item{sp_accumulation_over_time.pdf}{Species accumulation curves over time (site-specific and total).}
\item{sp_shared_among_sites.pdf}{Total number of species shared among sites across all dates.}
}
\description{
Some sort of a description
}
\details{

}
\note{
Not any \code{dataset} can be passed to this function. Some level of processing is almost always required. For this function to work well, replicate samples in a \code{dataset} observation should be summarized into a single value for a unique location_id, observation_datetime, taxon_id. Additionally, this function can only handle one variable_name and unit combination (sometimes a dataset will have multiple measurement types stacked in an observation table).
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
