% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_faceted_densities}
\alias{plot_faceted_densities}
\title{Faceted densities plot}
\usage{
plot_faceted_densities(tables, id, rank = NA, alpha = 1)
}
\arguments{
\item{tables}{(list of tbl_df, tbl, data.frame) A named list of ecocomDP tables.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{rank}{(string) The rank of taxa to plot. Defaults to NA}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Faceted densities plot
}
\examples{
tables <- ants_L1[[1]]$tables
id <- names(ants_L1)
rank <- "species"

plot_stacked_taxa_by_site(tables, id, rank)

}
