% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_sample_space_time}
\alias{plot_sample_space_time}
\title{Plot dates and times samples were collected or observations were made}
\usage{
plot_sample_space_time(data, id = NA_character_, alpha = 1)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, an ecocomDP observation table, or any table containing the columns present in an ecocomDP observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot dates and times samples were collected or observations were made
}
\examples{
\dontrun{
# plot ecocomDP formatted dataset
plot_sample_space_time(ants_L1)

# plot flattened ecocomDP dataset
plot_sample_space_time(flatten_data(ants_L1))

# plot ecocomDP observation table
plot_sample_space_time(ants_L1[[1]]$tables$observation)

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_sample_space_time()

# tidy syntax, filter data by site ID
ants_L1 \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    as.numeric(location_id) > 4) \%>\%
  plot_sample_space_time()
}

}
