% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community.R
\name{plot_sampling_effort}
\alias{plot_sampling_effort}
\title{Plot spatiotemporal sampling effort}
\usage{
plot_sampling_effort(dslong, id, path, alpha)
}
\arguments{
\item{dslong}{(data.frame) Long form of return from \code{format_for_comm_plots()}}

\item{id}{(character) Dataset id}

\item{path}{(character) Path to where plots will be written}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\value{
\item{sampling_effort.pdf}{Spatiotemporal sampling effort}
}
\description{
Plot spatiotemporal sampling effort
}
