% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_sp_accumulation_over_space}
\alias{plot_sp_accumulation_over_space}
\title{Plot species accumulation curve over space}
\usage{
plot_sp_accumulation_over_space(dslong, id, path, vunit)
}
\arguments{
\item{dslong}{(data.frame) Long form of return from \code{format_for_comm_plots()}}

\item{id}{(character) Dataset id}

\item{path}{(character) Path to where plots will be written}

\item{vunit}{(character) Unit of measurement variable}
}
\value{
\item{sp_accumulation_over_space.pdf}{Species accumulation curve over space}
}
\description{
Plot species accumulation curve over space
}
