% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community.R
\name{plot_sp_accumulation_over_time}
\alias{plot_sp_accumulation_over_time}
\title{Plot species accumulation curves over time (site-specific and total)}
\usage{
plot_sp_accumulation_over_time(dslong, id, path, vunit, alpha)
}
\arguments{
\item{dslong}{(data.frame) Long form of return from \code{format_for_comm_plots()}}

\item{id}{(character) Dataset id}

\item{path}{(character) Path to where plots will be written}

\item{vunit}{(character) Unit of measurement variable}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\value{
\item{sp_accumulation_over_time.pdf}{Species accumulation curves over time (site-specific and total)}
}
\description{
Plot species accumulation curves over time (site-specific and total)
}
