% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_accum_time}
\alias{plot_taxa_accum_time}
\title{Plot taxa accumulation through time}
\usage{
plot_taxa_accum_time(
  dataset = NULL,
  flat_data = NULL,
  observation = NULL,
  id = NA_character_,
  alpha = 1
)
}
\arguments{
\item{dataset}{(list) An ecocomDP formatted dataset}

\item{flat_data}{(tbl_df, tbl, data.frame) A flat data.frame created from an \code{ecocomDP} formatted dataset using \code{ecocomDP::flatten_data}.}

\item{observation}{(tbl_df, tbl, data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot taxa accumulation through time
}
\examples{
# plot ecocomDP formatted dataset
plot_taxa_accum_time(
  dataset = ants_L1)

# plot flattened ecocomDP dataset
plot_taxa_accum_time(
  flat_data = flatten_dataset(ants_L1))

# plot ecocomDP observation table
plot_taxa_accum_time(
  observation = ants_L1[[1]]$tables$observation)

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_dataset() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_accum_time(flat_data = .)
  
}
