% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_diversity}
\alias{plot_taxa_diversity}
\title{Plot diversity (taxa richness) through time}
\usage{
plot_taxa_diversity(
  data,
  id = NA_character_,
  time_window_size = "day",
  alpha = 1
)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, an ecocomDP observation table, or any table containing the columns present in an ecocomDP observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{time_window_size}{(character) Define the time window over which to aggregate observations for calculating richness}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot diversity (taxa richness) through time
}
\examples{
\dontrun{
# plot richness through time for ecocomDP formatted dataset by 
# observation date
plot_taxa_diversity(ants_L1)

# plot richness through time for ecocomDP formatted dataset by 
# aggregating observations within a year
plot_taxa_diversity(
  data = ants_L1,
  time_window_size = "year")

# plot richness through time for ecocomDP observation table
plot_taxa_diversity(ants_L1[[1]]$tables$observation)

# plot richness through time for flattened ecocomDP dataset 
plot_taxa_diversity(flatten_data(ants_L1))


# Using Tidy syntax:
# plot ecocomDP formatted dataset richness through time by 
# observation date
ants_L1 \%>\% plot_taxa_diversity()

ants_L1 \%>\% plot_taxa_diversity(time_window_size = "day")

# plot ecocomDP formatted dataset richness through time 
# aggregating observations within a month
ants_L1 \%>\% plot_taxa_diversity(time_window_size = "month")

# plot ecocomDP formatted dataset richness through time 
# aggregating observations within a year
ants_L1 \%>\% plot_taxa_diversity(time_window_size = "year")

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2007-01-01") \%>\%
  plot_taxa_diversity(
    time_window_size = "year")
}

}
