% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_occurrence_frequency}
\alias{plot_taxa_occurrence_frequency}
\title{Plot stacked taxa by site}
\usage{
plot_taxa_occurrence_frequency(
  dataset = NULL,
  cutoff = 0,
  rank = NA_character_,
  color_var = NA_character_,
  facet_var = NA_character_,
  alpha = 1
)
}
\arguments{
\item{dataset}{(list) An ecocomDP formatted dataset}

\item{cutoff}{(numeric) Defaults to 0.}

\item{rank}{(string) The rank of taxa to plot. Defaults to NA.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device.
}
\description{
Plot stacked taxa by site
}
\examples{
tables <- ants_L1[[1]]$tables
id <- names(ants_L1)
rank <- "species"

plot_stacked_taxa_by_site(tables, id, rank)

}
