% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_rank}
\alias{plot_taxa_rank}
\title{Plot taxa ranks}
\usage{
plot_taxa_rank(
  data,
  id = NA_character_,
  facet_var = NA_character_,
  facet_scales = "free",
  alpha = 1
)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, a flattened ecocomDP dataset, or any flat table containing the columns present in both the ecocomDP observation and taxon tables.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{facet_var}{(character) Name of column to use for faceting.}

\item{facet_scales}{(character) Should scales be free ("free", default value), fixed ("fixed"), or free in one dimension ("free_x", "free_y")?}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot taxa ranks
}
\examples{
\dontrun{
# plot ecocomDP formatted dataset
plot_taxa_rank(ants_L1)

# download and plot NEON macroinvertebrate data
my_dataset <- read_data(
  id = "neon.ecocomdp.20120.001.001",
  site= c('COMO','LECO'), 
  startdate = "2017-06",
  enddate = "2019-09",
  check.size = FALSE)

plot_taxa_rank(my_dataset)

# facet by location
plot_taxa_rank(
  data = my_dataset,
  facet_var = "location_id")

# plot flattened ecocomDP dataset
plot_taxa_rank(
  data = flatten_data(my_dataset),
  facet_var = "location_id")


# tidy syntax, filter data by date
my_dataset \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_rank()

# tidy syntax, filter data by site ID
my_dataset \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    grepl("COMO",location_id)) \%>\%
  plot_taxa_rank()
}

}
