% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_taxa_shared_sites}
\alias{plot_taxa_shared_sites}
\title{Plot number of taxa shared among sites}
\usage{
plot_taxa_shared_sites(observation, id)
}
\arguments{
\item{observation}{(data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}
}
\description{
Plot number of taxa shared among sites
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
