% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{pull_from_queue}
\alias{pull_from_queue}
\title{Pull next item from the ecocom-listener's queue}
\usage{
pull_from_queue(repository)
}
\arguments{
\item{repository}{(character) Repository hosting the ecocom-listener. Facilitates repository specific methods. Currently supported is: EDI.}
}
\value{
\item{index}{(integer) Index of item in queue. Is later used for removing the item from the queue.}
\item{package.id}{(character) Data package identifier}
\item{config.environment}{(character) Location of the \code{package.id} within a repository system. This variable is written to the Global Environment for use in \code{routine_handler()}.}
}
\description{
Pull next item from the ecocom-listener's queue
}
