% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data_edi}
\alias{read_data_edi}
\title{Read an ecocomDP dataset from EDI}
\usage{
read_data_edi(id)
}
\arguments{
\item{id}{(character) Data package identifier with revision number}
}
\value{
(list) Named list of data tables
}
\description{
Read an ecocomDP dataset from EDI
}
