% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_dir}
\alias{read_dir}
\title{Read L1 tables in path}
\usage{
read_dir(paths)
}
\arguments{
\item{paths}{(character) One or more paths to a directories}
}
\value{
(list) The L1 dataset object
}
\description{
Read L1 tables in path
}
