% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{Read EML metadata from a data repository}
\usage{
read_eml(package.id)
}
\arguments{
\item{package.id}{(character) Data package identifier}
}
\value{
(xml_document, xml_node) EML metadata
}
\description{
A wrapper function to repository specific read methods (the repository arg drives the logic).
}
\examples{

}
