% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_example_dataset}
\alias{read_example_dataset}
\title{Read example dataset from /data}
\usage{
read_example_dataset()
}
\value{
(list) Example dataset
}
\description{
Read example dataset from /data
}
