% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_example_L0_wide}
\alias{read_example_L0_wide}
\title{Read example L0_wide table}
\usage{
read_example_L0_wide()
}
\value{
(data.frame) Example L0_wide table
}
\description{
Read example L0_wide table
}
