% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data_edi.R
\name{read_table_dataset_summary}
\alias{read_table_dataset_summary}
\title{Read the ecocomDP dataset_summary table (from EDI)}
\usage{
read_table_dataset_summary(package.id)
}
\arguments{
\item{package.id}{(character) Data package identifier of an ecocomDP dataset in the EDI
data repository.}
}
\value{
(tibble) The dataset_summary table
}
\description{
Use this function to read the dataset_summary table of an ecocomDP data 
    package from the EDI Data Repository.
}
