% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_all_ecocomDP.R
\name{read_table_dataset_summary_neon}
\alias{read_table_dataset_summary_neon}
\title{Read the ecocomDP dataset_summary table (from NEON)}
\usage{
read_table_dataset_summary_neon(
      package.id
    )
}
\arguments{
\item{package.id}{(character) Data package identifier of an ecocomDP dataset in the EDI
data repository.}
}
\value{
(tibble) The dataset_summary table
}
\description{
Use this function to read the dataset_summary table of an ecocomDP data 
    package from the NEON Data Portal.
}
