% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{resolve_sci_taxa}
\alias{resolve_sci_taxa}
\title{Resolve scientific names to an authority}
\usage{
resolve_sci_taxa(x = NULL, data.sources, path = NULL)
}
\arguments{
\item{x}{(character) A vector of taxa names.}

\item{data.sources}{(numeric) An ordered vector of authority IDs to be queried (
\code{view_taxa_authorities()} lists currently supported authorities).
Taxa are resolved to authorities in the order listed. If an authority
and ID match cannot be made, then the next will be queried.}

\item{path}{(character) Path to directory containing taxa_map.csv. This tracks
provenance throughout the data cleaning process. Create this file with
\code{create_taxa_map()}.}
}
\value{
(data frame; taxa_map.csv) If using \code{x}, then a data frame
    containing the input taxa, accepted taxa name, rank, authority,
    authority ID, and score are returned. If using \code{path}, then an
    updated version of taxa_map.csv will be returned to \code{path} and
    a data frame of taxa_map.csv to the R environment.
}
\description{
Resolve taxa to preferred authorities and get associated IDs.
}
