% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{routine_handler}
\alias{routine_handler}
\title{Event notification workflow handler}
\usage{
routine_handler(config)
}
\arguments{
\item{config}{(character) Full path to config.R which contains global variables for parameterizing this function. Create config.R with \code{create_config()}}
}
\description{
Runs on a chron. Checks for unprocessed items in SQlite using webservices defined in ADD WEBSERVICE DEFINITIONS HERE and pops the longest waiting item into the appropriate routine. Currently supported are \code{update_L1()} or \code{update_L2_dwca()}.
}
