% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Search ecocomDP data}
\usage{
search(text, taxa, num_taxa, num_years, sd_years, area, boolean = "AND")
}
\arguments{
\item{text}{(character) Text to search in dataset titles, descriptions, and abstracts. Datasets matching any exact words or phrase will be returned. Is not case sensitive. Works with \code{boolean}}

\item{taxa}{(character) Taxanomic rank values to search on. The full taxonomic hierarchy of each taxa in a dataset is searchable for EDI (including common names) but not yet NEON, in which cases the lowest level rank value is searchable.}

\item{num_taxa}{(numeric) Minimum and maximum number of taxa the dataset should contain. Any datasets within this range will be returned.}

\item{num_years}{(numeric) Minimum and maximum number of years sampled the dataset should contain. Any datasets within this range will be returned.}

\item{sd_years}{(numeric) Minimum and maximum standard deviation between survey dates (in years). Any datasets within this range will be returned.}

\item{area}{(numeric) Bounding coordinates within which the data should originate. Accepted values are in decimal degrees and in the order: North, East, South, West. Any datasets with overlapping areas or contained points will be returned.}

\item{boolean}{(character) Boolean operator to use when searching \code{text} and 
\code{taxa}. Supported operators are: "AND", "OR". Default is "AND".}
}
\value{
(data.frame) Search results with these feilds:
    \itemize{
        \item source - Source from which the data originate. Currently 
        supported are "EDI" and "NEON".
        \item id - Identifier of dataset
        \item title - Title of dataset
        \item description - Description of dataset
        \item abstract - Abstract of dataset
        \item num_years - Number of years sampled
        \item sd_between_samples - Standard deviation between sampling 
        events in years.
        \item sites - Sites names or abbreviations (returned for NEON data
        sources but not EDI)
        \item url - URL to dataset
    }
}
\description{
Use this function to search across ecocomDP data made available by the 
Environmental Data Initiative (EDI) and the National Ecological Observatory
Network (NEON).
}
\details{
Currently, to accomodate multiple L1 versions of NEON data products, search results for a NEON L0 will also list all the L1 versions available for the match. This method is based on the assumption that the summary data among L1 versions is the same, which may need to be addressed in the future. A list of L0 and corresponding L1 identifiers are listed in /inst/L1_versions.txt. Each L1 version is accompanied by qualifying text that's appended to the title, abstract, and descriptions for comprehension of the differences among L1 versions.
}
\note{
This function may not work between 01:00 - 03:00 UTC due to regular maintenance of the EDI Data Repository. If you have reached this warning outside these hours then there may be an unexpected issue that will be resolved shortly. Please try again later.
    
    NEON taxa rank is lowest level identifiable. Higher ranks are not searchable.
}
\examples{
dontrun{
# Empty search returns all available data
View(search())

# Text searches titles, descriptions, and abstracts
View(search(text = "Lake"))

# Taxa searches any taxonomic rank of an organism
View(search(taxa = "Plantae"))

# Number of taxa searches the count of unique taxa in a dataset
View(search(num_taxa = c(0, 10)))

# Years searches the number of years the dataset covers
View(search(num_years = c(10, 20)))

# Standard deviation between surveys searches for a sampling frequency
View(search(sd_years = c(.25, 1)))

# Geographic area searches where data were collected
View(search(area = c(47.1, -86.7, 42.5, -92)))

# Boolean operators AND (default) & OR combine text and taxa search terms
View(search(text = c("Lake", "River")))
View(search(text = c("Lake", "River"), boolean = "OR"))
View(search(taxa = c("Plantae", "Animalia")))
View(search(taxa = c("Plantae", "Animalia"), boolean = "OR"))

# Use any combination of search fields to find the data you're looking for
r <- search(
  text = c("Lake", "River"),
  taxa = c("Plantae", "Animalia"),
  num_taxa = c(0, 10),
  num_years = c(10, 100),
  sd_years = c(.01, 100),
  area = c(47.1, -86.7, 42.5, -92),
  boolean = "OR")
View(r)
}
        
}
