% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send email}
\usage{
send_email(
  from,
  to,
  attachment,
  smtp.relay,
  relay.user,
  relay.user.pass,
  subject,
  msg
)
}
\arguments{
\item{from}{(character) Email address}

\item{to}{(character) Email address}

\item{attachment}{(character) Attachment file name with full path and file extension}

\item{smtp.relay}{(character) SMTP relay}

\item{relay.user}{(character) Relay user}

\item{relay.user.pass}{(character) Relay user password}

\item{subject}{(character) Subject line}

\item{msg}{(character) Message}
}
\description{
Send email
}
\details{
Works for Linux. May not work for other OS.
}
