% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{set_taxonomic_coverage}
\alias{set_taxonomic_coverage}
\title{Create the taxonomicCoverage EML node}
\usage{
set_taxonomic_coverage(sci_names)
}
\arguments{
\item{sci_names}{(list) Object returned by \code{get_classification()}.}
}
\value{
\item{list}{If \code{write.file = FALSE} an emld list object is returned
for use with the EML R Package.}
\item{.xml file}{If \code{write.file = TRUE} a .xml file is written to
\code{path}}.
}
\description{
Create the taxonomicCoverage EML node
}
