% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data_edi.R
\name{summarize_data_edi}
\alias{summarize_data_edi}
\title{Summarize supported EDI data packages}
\usage{
summarize_data_edi(path)
}
\arguments{
\item{path}{(character) Path the file will be saved to

This is a helper function for project maintainers. It summarizes supported
EDI data packages and saves it as an .rda data object to the /data 
directory of the ecocomDP R package, which is called upon by 
\code{search_data()} to quickly identify datasets of potential interest.}
}
\description{
Summarize supported EDI data packages
}
\examples{
\dontrun{
summarize_data_edi(path)
}
        
}
