% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data_neon.R
\name{summarize_data_neon}
\alias{summarize_data_neon}
\title{Summarize supported NEON data products}
\usage{
summarize_data_neon()
}
\description{
This is a helper function for project maintainers. It summarizes supported
NEON data products and saves it as an .rda data object to the /data 
directory of the ecocomDP R package, which is called upon by 
\code{search_data()} to quickly identify datasets of potential interest.
}
\examples{
\dontrun{
summarize_data_neon()
}
        
}
