% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable_name.R
\name{summarize_variable_name}
\alias{summarize_variable_name}
\title{Summarize variable name with function}
\usage{
summarize_variable_name(tables, varfun)
}
\arguments{
\item{tables}{(list) Data object returned by \code{read_data()[[1]]$tables}}

\item{varfun}{(named character) Variable to summarize (as name) and function to summarize with (as character; can be: "sum", "mean")}
}
\value{
\item{list}{\code{tables} with \code{names(varfun)} summarized by \code{varfun}}
}
\description{
Summarize variable name with function
}
\note{
Beware of replicate observations! These may have to be handled separately than by a blanket \code{sum()} or \code{mean()} call.
}
\examples{
\dontrun{
# Read dataset
d <- read_data("edi.124.4")
# Summarize the "Mass..g.m2." variable by mean()
summarize_variable_name(d[[1]]$tables, c(`Mass..g.m2.` = "mean"))
}

}
