% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_template_arguments.R
\name{to_docbook}
\alias{to_docbook}
\title{Convert to docbook}
\usage{
to_docbook(file = NULL)
}
\arguments{
\item{file}{(character) Full path to file to be read into docbook}
}
\value{
(xml) Docbook XML R object
}
\description{
This function is not exported from the EML R Package but useful here
}
