% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update an ecocomDP dataset}
\usage{
update_data(id, path)
}
\arguments{
\item{id}{(character) Identifier of L0 data package archived in the Environmental
Data Initiative Repository that has at least one revision that has been
converted to the ecocomDP and has a corresponding script that accords
with ecocomDP conversion best practices.}

\item{path}{(character) Path to which the revised ecocomDP tables, EML metadata, 
and message log will be written.}
}
\value{
\item{message log}{(.txt file) A report of the conversion process 
the results of the conversion process and highlighting any issues, 
or differences between L0 revisions that may have impact on the conversion 
accuracy and that helps maintainers quickly idenfity any manual changes 
that need to occur.}
\item{tables}{(delimited files) The L0 data in L1 formatted tables using the
same record delimiter as used in the L0.}
\item{metadata}{(.xml) Metadata in the EML format.}
}
\description{
A function for ecocomDP dataset maintainers that helps recreate the 
    L0 to L1 conversion when an L0 has been updated.
}
