% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dirty_bit.R
\name{update_dirty_bit}
\alias{update_dirty_bit}
\title{Update SQlite dirty bit}
\usage{
update_dirty_bit()
}
\value{
(character) "processed" if successful, "unprocessed" if unsuccessful
}
\description{
The dirty bit indicates whether an item is unprocessed or processed
}
\examples{

}
