% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_L2_dwca.R
\name{update_L2_dwca}
\alias{update_L2_dwca}
\title{Update L2 DwC-A from updated L1}
\usage{
update_L2_dwca(
  id.L1.newest,
  id.L2.next,
  core.name,
  path,
  url,
  user.id,
  user.pass
)
}
\arguments{
\item{id.L1.newest}{(character) Identifier of newest L1 data package}

\item{id.L2.next}{(character) Identifier of L1's new DwC-A child to be created by this function}

\item{core.name}{(character) The Darwin Core central table of the package. Can be: "event" (event core).}

\item{path}{(character) Directory to which L2 tables, meta.xml, and metadata will be written.}

\item{url}{(character) Publicly accessible URL to \code{path} for download by a data repository.}

\item{user.id}{(character) User identifier within a specified \code{repository}. This controls editing access in some \code{repository}.}

\item{user.pass}{(character) Password associated with \code{user.id} for repository upload.}
}
\description{
Updates an L2 DwC-A data package when it’s L1 parent data package has been updated. This function is a wrapper to several subroutines.
}
\details{
No comparisons of L1 newest and previous are required since they are both in a standardized format.
}
\note{
\code{user.id} and \code{user.pass} should be a set of master credentials within \code{repository}, otherwise issues at the evaluation/upload step may arise. Requires an L2 already exists in the \code{repository}. A pre-existing L2 must exist because the human has to decide if the L0 data already exists in GBIF or should not be uploaded to GBIF for other reasons.
}
\examples{

}
