% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{update_neon_summary_info}
\alias{update_neon_summary_info}
\title{Update NEON summary info}
\usage{
update_neon_summary_info(d)
}
\arguments{
\item{d}{(list) EDI and NEON dataset summary object}
}
\value{
(list) \code{d} with added info
}
\description{
This temporarily accomodates NEON's plan for multiple L1 versions and will have to be refactored to support other data sources adopting this use case. A permanent solution is to refactor the \code{summarize_data_neon()} and \code{summarize_data_edi()} functions to return the exact objects of interest.
}
\details{
Joins /inst/L1_versions.txt (a file holding version info) with \code{d} and does a little reformatting.
}
