% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_L2_dwca_to_gbif.R
\name{upload_L2_dwca_to_gbif}
\alias{upload_L2_dwca_to_gbif}
\title{Upload an L2 DwC-A to GBIF}
\usage{
upload_L2_dwca_to_gbif()
}
\value{

}
\description{
Upload an L2 DwC-A to GBIF
}
\examples{

}
