% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_repository.R
\name{upload_to_repository}
\alias{upload_to_repository}
\title{Upload data package to a repository}
\usage{
upload_to_repository(path, package.id, user.id, user.pass)
}
\arguments{
\item{path}{(character) Directory in which L1 tables, scripts, and metadata exist.}

\item{package.id}{(character) Identifier of data package to be uploaded}

\item{user.id}{(character) User identifier within a specified \code{repository}. This controls editing access in some \code{repository}.}

\item{user.pass}{(character) Password associated with \code{user.id} for repository upload.}
}
\value{
(character) Evaluation/upload summary
}
\description{
A wrapper function to repository specific upload methods.
}
\examples{

}
