% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{url_env}
\alias{url_env}
\title{Make URL for PASTA+ environment}
\usage{
url_env(environment)
}
\arguments{
\item{environment}{(character) Data repository environment to perform the evaluation in.
Can be: 'development', 'staging', 'production'.}
}
\description{
Create the URL suffix to the PASTA+ environment specified by the
    environment argument.
}
