% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Check if a dataset conforms to the ecocomDP model}
\usage{
validate(dataset = NULL, path = NULL)
}
\arguments{
\item{dataset}{(list) A dataset returned by \code{read()}.}

\item{path}{(character) A path to the directory in which the ecocomDP tables can be found.}
}
\value{
(character) If any checks fail, then a list of validation issues are returned along with a warning. If no issues are found then NULL is returned.
}
\description{
Check if a dataset conforms to the ecocomDP model
}
\details{
Validation checks:
   \itemize{
       \item{File names - File names are the ecocomDP table names.}
       \item{Table presence - Required tables are present.}
       \item{Column names - Column names of all tables match the model.}
       \item{Column presence - Required columns are present.}
       \item{Column classes - Column classes match the model specification.}
       \item{Datetime format - Date and time formats follow the model specification.}
       \item{Primary keys - Primary keys of tables are unique.}
       \item{Composite keys - Composite keys (unique constraints) of each table are unique.}
       \item{Referential integrity - Foreign keys have a corresponding primary key.}
       \item{Coordinate format - Values are in decimal degree format.}
       \item{Coordinate range - Values are within -90 to 90 and -180 to 180.}
       \item{Elevation - Values are less than Mount Everest (8848 m) and greater than Mariana Trench (-10984 m).}
       \item{Variable mapping - variable_name is in table_name.}
   }
}
\note{
This function is used by ecocomDP creators (to ensure what has been created is valid), maintainers (to improve the quality of archived ecocomDP datasets), and users (to ensure the data being used is free of error).
}
\examples{
# Validate a set of files
r <- validate(path = system.file("/data", package = "ecocomDP"))

# Validate a list object
CHANGE EXAMPLE TO VALIDATE A LIST OBJECT NOT CONSTRUCTED BY read() ... BECAUSE THIS FUNC ALREADY VALIDATES (CIRCULAR EXAMPLE) ... IS THIS REALLY A USE CASE?
d <- read(from = system.file("/data", package = "ecocomDP"))
r <- validate(dataset = d)

}
