% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_column_classes}
\alias{validate_column_classes}
\title{Check for valid column classes}
\usage{
validate_column_classes(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
If column classes are not formmated to specification, then an error 
    is returned.
}
\description{
Check for valid column classes
}
