% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_column_presence.R
\name{validate_column_presence}
\alias{validate_column_presence}
\title{Validate column presence}
\usage{
validate_column_presence(data.path, criteria)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}

\item{criteria}{A data frame of the validation criteria located in 
/inst/validation_criteria.txt.}
}
\value{
A validation report printed in the console window of RStudio.
}
\description{
This function ensures that your ecocomDP (L1) tables contain required 
    columns.
}
\details{
The full suite of L1 validation checks are performed by the 
   \code{validate_ecocomDP} function. The sequence of checks performed in
   \code{validate_ecocomDO} are not random, rather some checks are dependent
   upon others.
}
