% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_composite_keys}
\alias{validate_composite_keys}
\title{Check composite column uniqueness}
\usage{
validate_composite_keys(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
If composite keys are not unique an error is returned.
}
\description{
Check composite column uniqueness
}
