% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate a dataset against the ecocomDP}
\usage{
validate_data(dataset = NULL, data.path = NULL)
}
\arguments{
\item{dataset}{(list of data frames) A single dataset return from \code{read_data()}.}

\item{data.path}{(character) The path to the directory containing ecocomDP tables.}
}
\value{
(character) If issues are found, then a vector of character strings 
    describing validation issues is returned along with a warning. If no 
    issues are found then a NULL is returned.
}
\description{
Use this function to verify a dataset conforms to the ecocomDP.
}
\details{
Validation checks:
   \itemize{
       \item{File names - When using \code{data.path} the data are read from
       files and must follow an expected pattern, namely 
       \emph{studyName_ecocomDPTableName.ext} (e.g. 
       \emph{gleon_chloride_observation.csv}).}
       \item{Table presence - Some tables are required, others are not.}
       \item{Column names - Column names must follow specification.}
       \item{Column presence - Some columns are required, others are not.}
       \item{Column classes - Column classes must follow specification.}
       \item{Datetime format - Date and datetime formats must follow 
       specification.}
       \item{Primary keys - Primary keys of tables must be valid.}
       \item{Composite keys - Composite keys of tables must be valid.}
       \item{Referential integrity - Referential integrity among tables must 
       be valid.}
   }
}
\note{
This function is used by ecocomDP creators (to ensure what has been 
    created is valid), maintainers (to improve the quality of archived
    ecocomDP datasets), and users (to ensure the data being used is free of
    model error).
}
\examples{
# Validate a set of files
r <- validate_data(data.path = system.file("/data", package = "ecocomDP"))

# Validate a list object
CHANGE EXAMPLE TO VALIDATE A LIST OBJECT NOT CONSTRUCTED BY read_data() ... BECAUSE THIS FUNC ALREADY VALIDATES (CIRCULAR EXAMPLE) ... IS THIS REALLY A USE CASE?
d <- read_data(from.files = system.file("/data", package = "ecocomDP"))
r <- validate_data(dataset = d)
        
}
