% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_dataset_structure}
\alias{validate_dataset_structure}
\title{Validate dataset structure}
\usage{
validate_dataset_structure(dataset)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}
}
\description{
Validate dataset structure
}
\details{
Returns an error if \code{dataset} is malformed
}
