% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_datetime}
\alias{validate_datetime}
\title{Check for valid date and time formats}
\usage{
validate_datetime(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
If date and time data are not formmated to specification, then an error 
    is returned.
}
\description{
Check for valid date and time formats
}
