% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_ecocomDP}
\alias{validate_ecocomDP}
\title{Validate ecocomDP tables}
\usage{
validate_ecocomDP(data.path, sep)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}

\item{datetime.format.string}{A character string specifying the datetime format string used in your
L1 tables.}
}
\value{
A validation report printed in the console window of RStudio. The 
    validation checks run until an error is encountered. You must address 
    errors before continuing to the next check. Once all validation checks 
    have successfully completed, you will be notified with a congratulatory 
    message.
}
\description{
Once you've formatted a dataset to the ecocomDP (L1) you will need to 
    check that it complies with the data pattern standard. This process 
    ensures your L1 is valid and can be combined with other L1 with the 
    \code{aggregate_ecocomDP} function.
}
\details{
Run this function after creating a L1 and before making EML for it.

   Validation checks performed by this function:
   \itemize{
       \item \strong{Table names} Table names must follow the ecocomDP 
       naming convention (i.e. \emph{studyName_ecocomDPTableName.ext}, e.g. 
       \emph{gleon_chloride_observation.csv}). 
       \item \strong{Required tables} Some L1 tables are 
       required, others are not.
       \item \strong{Column names} Column names must be those specified by
       the data pattern.
       \item \strong{Required columns} Some columns are 
       required, others are not.
       \item \strong{Column classes} Column classes must match the 
       ecocomDP specification.
       \item \strong{Datetime format string} Datetime values should match
       the input datetime.format.string parameter.
   }
}
