% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_elevation}
\alias{validate_elevation}
\title{Check elevation}
\usage{
validate_elevation(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If elevation values are greater than Mount Everest (8848 m) 
    or are less than the Mariana Trench (-10984 m), then a message is returned.
}
\description{
Check elevation
}
