% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fields.R
\name{validate_fields}
\alias{validate_fields}
\title{Validate table fields}
\usage{
validate_fields(path = "", data.files = c("data.file.1", "data.file.2", "etc."))
}
\arguments{
\item{path}{A character string specifying a path to the dataset working directory.}

\item{data.files}{A list of character strings specifying the names of the data files
of your dataset. It is not necessary to include the file extension.}
}
\value{
If incongruence is found an error is returned.
}
\description{
This function checks for congruence in the number of fields among rows 
    of a data table.
}
