% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_file_names}
\alias{validate_file_names}
\title{Validate file names}
\usage{
validate_file_names(path, data.files)
}
\arguments{
\item{path}{(character) A character string specifying a path to the dataset working 
directory.}

\item{data.files}{A list of character strings specifying the names of the data files of 
your dataset.}
}
\value{
A warning message if the data files don't exist at path, and which of
    the input data files are missing.
    
    The full names of files listed in the data.files argument.
}
\description{
Identify whether input data file names exist in the specified directory.
}
