% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_latitude_longitude_format}
\alias{validate_latitude_longitude_format}
\title{Check latitude and longitude format}
\usage{
validate_latitude_longitude_format(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If latitude and longitude values in the location table are 
    not in decimal degree format, then a message is returned.
}
\description{
Check latitude and longitude format
}
