% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_latitude_longitude_range}
\alias{validate_latitude_longitude_range}
\title{Check latitude and longitude range}
\usage{
validate_latitude_longitude_range(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If latitude and longitude values in the location table are 
    outside of the expected ranges -90 to 90 and -180 to 180 respectively,
    then a message is returned.
}
\description{
Check latitude and longitude range
}
