% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_path.R
\name{validate_path}
\alias{validate_path}
\title{Validate path}
\usage{
validate_path(path = "")
}
\arguments{
\item{path}{A character string specifying a path to the dataset working directory.}
}
\value{
A warning message if the path leads to a non-existant directory.
}
\description{
Use \code{dir.exists} to determine whether the input path is valid and 
    returns an error message if not.
}
