% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_primary_keys.R
\name{validate_primary_keys}
\alias{validate_primary_keys}
\title{validate_primary_keys}
\usage{
validate_primary_keys(x, col)
}
\arguments{
\item{x}{(data frame) The table to be checked.}

\item{col}{(character) Name of the column containing primary keys.}
}
\value{
An error is issued if primary keys are not unique.
}
\description{
Checks that primary keys of a table are unique and issues an error if
    otherwise.
}
