% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_referential_integrity}
\alias{validate_referential_integrity}
\title{Check that referential integrity of ecocomDP tables is valid}
\usage{
validate_referential_integrity(tables, data.path, criteria)
}
\arguments{
\item{tables}{(character) Names of valid L1 tables in your dataset working directory.
Get valid table names with `validate_table_names`.}

\item{data.list}{(list of data frames) A list of of data frames. Use this argument if
tables are read into the R environment. The list should be of the following 
structure:
\itemize{
    \item \strong{"list name"} Name of the list object. This can be anything.
    \itemize{
        \item \strong{"L1 table name"} i.e. observation, location, taxon, etc.
        \itemize{
            \item \strong{"data"} Data frame representation of the L1 table
            \item \strong{"fname"} File name the data frame was created from
        }
    }
}}

\item{criteria}{(data frame) Validation criteria located at 
/inst/validation_criteria.txt.}
}
\value{
If foreign keys do not match primary keys, then an error is returned.
}
\description{
Check that foreign keys match primary keys.
}
