% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_presence.R
\name{validate_table_presence}
\alias{validate_table_presence}
\title{Validate presence of ecocomDP tables}
\usage{
validate_table_presence(data.path)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}
}
\value{
A validation report printed in the RStudio console window. The 
    validation check runs until an error is encountered. You must address 
    errors of one table before the next table is checked. Once all 
    tables have been checked you will be notified with a congratulatory 
    message.
}
\description{
This function ensures that your ecocomDP (L1) tables include all 
    required tables (i.e. \emph{location, taxon, observation, 
    dataset_summary}).
}
\details{
The full suite of L1 validation checks are performed by the 
   \code{validate_ecocomDP} function. The sequence of checks performed in
   \code{validate_ecocomDO} are not random, rather some checks are dependent
   upon others.
}
