% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_validate.R
\name{validate_taxonomic_coverage}
\alias{validate_taxonomic_coverage}
\title{Validate the taxonomic_coverage template}
\usage{
validate_taxonomic_coverage(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template 
    grouped by "required" and "optional" criteria.}
    \item{x}{(named list) If "required" issues are found, then the template 
    is removed.}
}
\description{
Validate the taxonomic_coverage template
}
\details{
Checks performed by this function:
    \itemize{
        \item{Template column names}
        \item{Taxonomic authority}
        \item{Complete entries}
    }
    
    Manipulations performed by this function:
    \itemize{
        \item{Use raw names when a resolved name is missing. This ensures
        the taxon is not ommitted from the taxonomic coverage metadata.}
    }
}
