% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_validate.R
\name{validate_taxonomic_coverage_column_names}
\alias{validate_taxonomic_coverage_column_names}
\title{Check column names of taxonomic coverage template}
\usage{
validate_taxonomic_coverage_column_names(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{error}{If column names are invalid}
    \item{NULL}{If no issues were found}
}
\description{
Check column names of taxonomic coverage template
}
\details{
Adds optional columns when missing so call to taxonomyCleanr::make_taxonomicCoverage() works.
}
