% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAL_validate.R
\name{validate_taxonomic_coverage_completeness}
\alias{validate_taxonomic_coverage_completeness}
\title{Check completeness of the taxonomic coverage template}
\usage{
validate_taxonomic_coverage_completeness(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{error}{If column names are invalid}
    \item{NULL}{If no issues were found}
}
\description{
Check completeness of the taxonomic coverage template
}
