% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_variable_mapping}
\alias{validate_variable_mapping}
\title{Check variable_mapping}
\usage{
validate_variable_mapping(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If variable_name is not in table_name, then a message is returned.
}
\description{
Check variable_mapping
}
