% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_all_ecocomDP.R
\name{view_all_ecocomDP}
\alias{view_all_ecocomDP}
\title{View a summary of all ecocomDP datasets}
\usage{
view_all_ecocomDP()
}
\value{
A data frame with fields:
    \itemize{
        \item \strong{package_id}: Data package ID
        \item \strong{title}: Data package title
        \item \strong{original_package_id}: The parent data package ID
        \item \strong{length_of_survey_years}: The number of years surveyed
        \item \strong{number_of_years_sampled}: Number of years sampled
        \item \strong{std_dev_interval_betw_years}: Standard deviation interval between years
        \item \strong{max_num_taxa}: Number of unique taxa observed
        \item \strong{geo_extent_bounding_box_m2}: Area (square meters) over which samples were taken.
    }
}
\description{
View a summary of all ecocomDP datasets.
}
