% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_annotation_dictionary.R
\name{view_annotation_dictionary}
\alias{view_annotation_dictionary}
\title{Browse ecocomDP's dictionary for annotating datasets and attributes}
\usage{
view_annotation_dictionary()
}
\description{
View the collection of dataset- and attribute-level annotations
    from existing ecocomDP datasets.
}
\details{
Use the search field to find the annotation terms and URIs.
}
\examples{
\dontrun{
view_annotation_dictionary()
}
    
}
