% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_landing_page.R
\name{view_landing_page}
\alias{view_landing_page}
\title{view_landing_page}
\usage{
view_landing_page(id)
}
\arguments{
\item{id}{(integer) A single integer or vector of integers corresponding with 
data package IDs listed in the data frame created by 
`list_all_ecocomDP`.}
}
\value{
The data package landing page.
}
\description{
View human readable version of a data package EML record.
}
