% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{view_model_descriptions}
\alias{view_model_descriptions}
\title{View ecocomDP table descriptions and requirements}
\usage{
view_model_descriptions()
}
\description{
View ecocomDP table descriptions and requirements
}
\examples{
\dontrun{
view_model_descriptions()
}

}
