% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyCleanr.R
\name{view_taxa_authorities}
\alias{view_taxa_authorities}
\title{View taxonomic authorities}
\usage{
view_taxa_authorities()
}
\value{
(data frame) Taxonomic authorities and corresponding identifiers
    supported by `resolve_taxa` and `resolve_common`.
}
\description{
List taxonomic authorities supported by taxonomic funcs.
}
\details{
View taxonomic authorities supported by `resolve_taxa` and
    `resolve_common`.
}
