% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_unit_dictionary.R
\name{view_unit_dictionary}
\alias{view_unit_dictionary}
\title{View the standard unit dictionary}
\usage{
view_unit_dictionary()
}
\description{
View the standard unit dictionary of the Ecological Metadata Language
    schema in the RStudio source window.
}
\details{
Use the search field to find the unit of interest.
}
