% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vocab_resolve_terms}
\alias{vocab_resolve_terms}
\title{Resolve terms to a controlled vocabulary}
\usage{
vocab_resolve_terms(x, cv, messages = FALSE, interactive = FALSE)
}
\arguments{
\item{x}{(character) Term(s) to resolve to a controlled vocabulary. Can be a 
vector of terms.}

\item{cv}{(character) A controlled vocabulary to search. Valid options are:
\itemize{
    \item{lter} - The LTER Controlled Vocabulary (http://vocab.lternet.edu/vocab/vocab/index.php)
}}

\item{messages}{(logical) Display diagnostic messages, e.g. alternative spelling options.}

\item{interactive}{(logical) Query user to select from alternative terms and returns back
selection.}
}
\value{
(character) Controlled vocabulary names corresponding to successfully
    resolved terms.
}
\description{
Resolve terms to a controlled vocabulary.
}
