% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ecocomDP_tables.R
\name{write_ecocomDP_tables}
\alias{write_ecocomDP_tables}
\title{Write ecocomDP tables to file}
\usage{
write_ecocomDP_tables(path, sep,  observation, location, taxon, 
    dataset_summary, observation_ancillary = NULL, 
    location_ancillary = NULL, taxon_ancillary = NULL, 
    variable_mapping = NULL)
}
\arguments{
\item{path}{(character) Directory to which the files will be written.}

\item{sep}{(character) Field delimiter used in the ecocomDP tables. Valid options 
are "," or "\\t".}

\item{study.name}{(character) Name to be prefixed to each ecocomDP table name (e.g. 
'bird_survey_').}

\item{observation}{(data frame) observation table}

\item{location}{(data frame) location table}

\item{taxon}{(data frame) taxon table}

\item{dataset_summary}{(data frame) dataset_summary table}

\item{observation_ancillary}{(data frame) observation_ancillary table}

\item{location_ancillary}{(data frame) location_ancillary table}

\item{taxon_ancillary}{(data frame) taxon_ancillary table}

\item{variable_mapping}{(data frame) variable_mapping table}
}
\value{
ecocomDP tables written to file.
}
\description{
Write ecocomDP tables to file
}
