% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_tables}
\alias{write_tables}
\title{Write ecocomDP tables to file}
\usage{
write_tables(
  path,
  sep = ",",
  observation = NULL,
  location = NULL,
  taxon = NULL,
  dataset_summary = NULL,
  observation_ancillary = NULL,
  location_ancillary = NULL,
  taxon_ancillary = NULL,
  variable_mapping = NULL
)
}
\arguments{
\item{path}{(character) A path to the directory in which the files will be written.}

\item{sep}{(character) Field delimiter to use when writing files. Default is comma.}

\item{observation}{(data.frame) The observation table.}

\item{location}{(data.frame) The location table.}

\item{taxon}{(data.frame) The taxon table.}

\item{dataset_summary}{(data.frame) The dataset_summary table.}

\item{observation_ancillary}{(data.frame) The observation_ancillary table.}

\item{location_ancillary}{(data.frame) The location_ancillary table.}

\item{taxon_ancillary}{(data.frame) The taxon_ancillary table.}

\item{variable_mapping}{(data.frame) The variable_mapping table.}
}
\value{
ecocomDP tables as \code{sep} delimited files
}
\description{
Write ecocomDP tables to file
}
\examples{
\dontrun{

}

}
