% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{xml_val}
\alias{xml_val}
\title{Get XML values}
\usage{
xml_val(nodeset, xpath)
}
\arguments{
\item{nodeset}{(xml_node/xml_nodeset) Nodeset}

\item{xpath}{(character) xpath}
}
\value{
(character) Value of \code{xpath} within \code{nodeset}. Returns "" if returned character string has length = 1.
}
\description{
Get XML values
}
\details{
Simplifies code by wrapping \code{cnvmt(xml2::xml_text(xml2::xml_find_all(...), trim = T))}
}
