% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ecocomDP.R
\name{aggregate_ecocomDP}
\alias{aggregate_ecocomDP}
\title{Aggregate ecocomDP data packages}
\usage{
aggregate_ecocomDP(
      package.id, 
      path = NULL
    )
}
\arguments{
\item{package.id}{(character) IDs of data packages to aggregate. Use 
\code{view_all_ecocomDP} to get package IDs of available ecocomDP 
datasets. Previous data package revisions are valid. It's OK to if only
one data package is entered.}

\item{path}{(character; optional) Path to the directory in which the aggregated 
ecocomDP will be written.}

\item{neon.sites}{(character) NEON sites to use the data of. THIS ARGUMENT IS A WIP.}
}
\value{
\itemize{
        \item{(list) A named list of tibbles (one for each ecocomDP table)
        containing data from aggregated \code{package.id}s}
        \item{(.csv files) .csv files (one for each ecocomDP table) 
        containing data from aggregated \code{package.id}s}
    }
}
\description{
Aggregate ecocomDP data packages.
}
