% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ecocomDP.R
\name{aggregate_ecocomDP}
\alias{aggregate_ecocomDP}
\title{aggregate_ecocomDP}
\usage{
view_all_ecocomDP()
}
\value{
A data frame with fields:
    \itemize{
        \item package_id: Data package ID
        \item title: Data package title
        \item original_package_id: The parent data package ID
        \item length_of_survey_years: The number of years surveyed
        \item number_of_years_sampled: Number of years sampled
        \item std_dev_interval_betw_years: Standard deviation interval between years
        \item max_num_taxa: Number of unique taxa observed
        \item geo_extent_bounding_box_m2: Area (square meters) over which samples were taken.
    }
}
\description{
View list of all available ecocomDP with high level metadata.
}
