% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_dictionary.R
\name{annotation_dictionary}
\alias{annotation_dictionary}
\title{Annotations of published data}
\usage{
annotation_dictionary()
}
\description{
View the collection of dataset- and attribute-level annotations
    from existing ecocomDP datasets.
}
\details{
Use the search field to find the annotation terms and URIs.
}
\examples{
\dontrun{
View(annotation_dictionary())
}
    
}
