% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ants_L0_flat}
\alias{ants_L0_flat}
\title{Joined and flat version of EDI data package knb-lter-hfr.118.32}
\format{
A data frame with 2931 rows and 45 variables:
\describe{
  \item{datetime}{Sample dates}
  \item{block}{Sample location}
  ...
}
}
\source{
\url{https://portal.edirepository.org/nis/mapbrowse?scope=knb-lter-hfr&identifier=118&revision=32}
}
\usage{
ants_L0_flat
}
\description{
A fully joined and flat version of EDI data package knb-lter-hfr.118.32 (Ant Assemblages in Hemlock Removal Experiment at Harvard Forest since 2003) with all relevant ecocomDP L1 identifiers and content added. Use this dataset as an input to the \code{L0_flat} argument of the "create" functions.
}
\keyword{datasets}
