% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary.R
\name{calc_length_of_survey_years}
\alias{calc_length_of_survey_years}
\title{Calculate length_of_survey_years for the dataset_summary table}
\usage{
calc_length_of_survey_years(dates)
}
\arguments{
\item{dates}{(Date) Dates from the L0 source dataset encompassing the entire study duration.}
}
\value{
(numeric) Number of years the study has been ongoing.
}
\description{
Calculate length_of_survey_years for the dataset_summary table
}
