% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coerce_table_classes}
\alias{coerce_table_classes}
\title{Coerce table classes to ecocomDP specifications}
\usage{
coerce_table_classes(tbl, name)
}
\arguments{
\item{tbl}{(data.frame) Table to coerce}

\item{name}{(character) Table name}
}
\value{
(data.frame) \code{tbl} with column classes coerced to ecocomDP model specifications
}
\description{
Coerce table classes to ecocomDP specifications
}
\details{
Datetime columns are not coerced. These are unchanged from the input class.
}
