% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dwca.R
\name{convert_to_dwca}
\alias{convert_to_dwca}
\title{Convert ecocomDP to Darwin Core Archive}
\usage{
convert_to_dwca(
  path,
  core_name,
  source_id,
  derived_id,
  url = NULL,
  user_id,
  user_domain
)
}
\arguments{
\item{path}{(character) Path to which the DwC-A data objects and EML will be written.}

\item{core_name}{(character) The central table of the Darwin Core package being created. Can be: "event" (event core). Occurrence core is not yet supported.}

\item{source_id}{(character) Identifier of an ecocomDP data package published in a supported repository. Currently, the EDI Data Repository is supported.}

\item{derived_id}{(character) Identifier of the DwC-A dataset being created.}

\item{url}{(character) URL to the publicly accessible directory containing DwC-A tables and meta.xml. This argument supports direct download of the data entities by a data repository and is used within the scope of the ecocomDP project for automated revision and publication of datasets.}

\item{user_id}{(character) Identifier of user account associated with the data repository in which this ecocomDP dataset will be archived. Only \code{user_id} from the EDI is currently supported.}

\item{user_domain}{(character) Domain of the \code{user_id}. Only "EDI" is currently supported. If more than one, then supply as a vector of character strings in the same order as \code{user_id}.}
}
\value{
DwC-A tables, meta.xml, and corresponding EML metadata.
}
\description{
Convert ecocomDP to Darwin Core Archive
}
\details{
Reads in an ecocomDP dataset from a supported repository and converts it to a DwC-A package.
}
\examples{

}
