% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary.R
\name{create_dataset_summary}
\alias{create_dataset_summary}
\title{Make dataset_summary table}
\usage{
create_dataset_summary(
  parent.package.id,
  child.package.id,
  sample.dates,
  taxon.table,
  north,
  east,
  south,
  west,
  environment = "production"
)
}
\arguments{
\item{parent.package.id}{(character) Parent data package identifier composed of scope, 
identifier, and revision (e.g. 'edi.101.1').}

\item{child.package.id}{(character) Child data package identifier composed of scope, 
identifier, and revision (e.g. 'edi.101.1').}

\item{sample.dates}{(character) Sample dates from which to calculate date fields of the
dataset_summary table.}

\item{taxon.table}{(data frame) The ecocomDP taxon table.}

\item{north}{(numeric) North bounding coordinate of the study area in decimal 
degrees with values north of the equator being positive.}

\item{east}{(numeric) East bounding coordinate of the study area in decimal 
degrees with values west of the prime meridian being negative.}

\item{south}{(numeric) South bounding coordinate of the study area in decimal 
degrees with values south of the equator being negative.}

\item{west}{(numeric) West bounding coordinate of the study area in decimal 
degrees with values west of the equator being negative.}

\item{evironment}{(character) EDI Data Repository environment in which the L0 exists, and which will be used to get the geographic bounding area from the EML metadata. Default is "production".}
}
\value{
(data frame) A data frame of the dataset_summary table
}
\description{
Make the ecocomDP dataset_summary table.
}
