% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eml.R
\name{create_eml}
\alias{create_eml}
\title{Make EML metadata for ecocomDP tables}
\usage{
create_eml(
  path,
  source_id,
  derived_id,
  script,
  script_description,
  is_about = NULL,
  contact,
  user_id,
  user_domain,
  basis_of_record = NULL,
  url = NULL
)
}
\arguments{
\item{path}{(character) Path to the directory containing ecocomDP data tables, conversion scripts, and where EML metadata will be written.}

\item{source_id}{(character) Parent data package ID (e.g. "knb-lter-hfr.118.28"). Only 
EDI Data Repository package IDs are currently supported.}

\item{derived_id}{(character) Child data package ID (e.g. "edi.53.1") to be uploaded to 
EDI Data Repository.}

\item{script}{(character) Names of scripts used to convert and repackage the source 
data into the ecocomDP.}

\item{script_description}{(character) A description for each object listed under \code{script}.}

\item{is_about}{(named character) Dataset level annotations describing what this dataset "is about" (L0 keywords converted to object label + URI pairs) describing what the L1 dataset is about. (e.g. \code{is_about = c('level of ecological disturbance' = "http://purl.dataone.org/odo/ECSO_00002588",' type of ecological disturbance' = "http://purl.dataone.org/odo/ECSO_00002589")}).}

\item{contact}{(data frame) Contact information of this ecocomDP creator.
The data frame must have these columns:
\itemize{
    \item{givenName}
    \item{surName}
    \item{organizationName}
    \item{electronicMailAddress}
}}

\item{user_id}{(character) ID(s) of user account(s) associated with the data repository
in which this ecocomDP data package will be archived. Only EDI Data 
Repository IDs are currently supported.}

\item{user_domain}{(character) Domain of the \code{user_id}(s). Only "EDI" and "LTER" are 
currently supported. If more than one, then supply as a vector of 
character strings in the same order as the corresponding \code{user_id}.}

\item{basis_of_record}{(character) To create a Darwin Core record from this ecocomDP using
\code{L1_to_L2_DwCA()} then the Darwin Core property 
\href{basisOfRecord}{https://dwc.tdwg.org/terms/#dwc:basisOfRecord}
requires definition as 
\href{HumanObservation}{http://rs.tdwg.org/dwc/terms/HumanObservation} or 
\href{MachineObservation}{http://rs.tdwg.org/dwc/terms/MachineObservation}.}

\item{url}{(character) URL to the publicly accessible directory containing ecocomDP data tables, conversion scripts, and EML metadata. This argument supports direct download of the data entities by a data repository and is used within the scope of the ecocomDP project for automated revision and upload of ecocomDP data packages and derived products.}
}
\value{
An EML metadata record for the ecocomDP dataset defined by the arguments
    to this function.
}
\description{
Run this function AFTER you have validated your ecocomDP tables with 
\code{validate()}.
}
\details{
This function creates an EML record for an Ecological Community Data 
    Pattern (ecocomDP) combining metadata from the parent data package and
    boiler-plate metadata describing the ecocomDP tables. Changes to the 
    parent EML include:
    \itemize{
        \item \strong{<access>} Adds the \code{user_id} to the list of 
        principals granted read and write access to the ecocomDP data 
        package this EML describes.
        \item \strong{<title>} Adds a note that this is a derived data 
        package in the ecocomDP format.
        \item \strong{<pubDate>} Adds the date this EML was created.
        \item \strong{<abstract>} Adds a note that this is a derived data 
        package in the ecocomDP format.
        \item \strong{<keywordSet} Adds the "ecocomDP" keyword to enable
        search and discovery of all ecocomDP data packages in the EDI Data
        Repository, and 7 terms from the LTER Controlled vocabulary:
        "communities", "community composition", "community dynamics", 
        "community patterns", "species composition", "species diversity",
        and "species richness". Darwin Core Terms listed under \code{
        basis_of_record} are listed and used by \code{L1_to_L2_DwCA()} 
        to create a Darwin Core Archive of this ecocomDP data package.
        \item \strong{<intellectualRights>} Keeps intact the intellectual
        rights license of the parent data package, or replaces it with
        "CCO" (https://creativecommons.org/publicdomain/zero/1.0/legalcode).
        \item \strong{<taxonomicCoverage>} Updates the taxonomic coverage 
        element with data supplied in the taxon table of the ecocomDP.
        \item \strong{<contact>} Adds contact information of the ecocomDP
        creator to the list of contacts in the parent data package EML.
        \item \strong{<methodStep>} Adds a note that this data package was
        created by the scripts listed under the \code{script} argument,
        and adds provenance metadata noting that this is a derived data 
        and describing where the parent data package can be accessed.
        \item \strong{<dataTables>} Replaces the parent data package data
        tables metadata with boiler-plate metadata for the ecocomDP tables.
        \item \strong{<otherEntity>} Describes scripts listed in the 
        \code{script} and \code{script_description} arguments. Any other
        entities listed in the parent EML are removed.
    }
    
    CURRENTLY ONLY WORKS FOR EDI DATA REPOSITORY
    
    TAXA RESOLVED USING taxonomyCleanr METHODS EXPORTED FROM THIS PACKAGE ARE, WILL HAVE THEIR RANKS EXPANDED CREATE_EML(). ALTERNATIVELY COULD DO IT YOUR SELF WITH THESE CONTROLLED TERMS THAT NEED TO BE FOLLOWED.
}
\examples{
\dontrun{

# Set path to ecocomDP tables and scripts
path <- "/Users/csmith/Desktop/ecocomDP"

# Validate ecocomDP tables
validate(path)

# Create contact information for the ecocomDP creator
additional_contact <- data.frame(
  givenName = 'Colin',
  surName = 'Smith',
  organizationName = 'Environmental Data Initiative',
  electronicMailAddress = 'csmith@wisc.edu',
  stringsAsFactors = FALSE
)

# Create EML
create_eml(
  path = path,
  source_id = "knb-lter-sev.29.12",
  derived_id = "edi.101.5",
  script = c(
    "convert_sev29_to_ecocomDP.R", 
    "package_edi_333.R"
  ),
  script_description = c(
    "R script for creating the ecocomDP tables.",
    "R script for creating the ecocomDP EML"
  ),
  contact = additional_contact,
  user_id = 'csmith',
  user_domain = 'LTER',
)
}

}
