% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_location_ancillary.R
\name{create_location_ancillary}
\alias{create_location_ancillary}
\title{Create the location_ancillary table}
\usage{
create_location_ancillary(
  L0_flat,
  location_id,
  datetime = NULL,
  variable_name,
  unit = NULL
)
}
\arguments{
\item{L0_flat}{(data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{location_id}{(character) Column in \code{L0_flat} containing the identifier assigned to each unique location at the observation level.}

\item{datetime}{(character) An optional column in \code{L0_flat} containing the date, and if applicable time, of ancillary location data following the ISO-8601 standard format (i.e. YYYY-MM-DD hh:mm:ss).}

\item{variable_name}{(character) Columns in \code{L0_flat} containing the ancillary location data.}

\item{unit}{(character) An optional column in \code{L0_flat} containing the units of each \code{variable_name} following the column naming convention: unit_<variable_name> (e.g. "unit_depth").}
}
\value{
(data.frame) The location_ancillary table.
}
\description{
Create the location_ancillary table
}
\details{
"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" ecocomDP tables can be consistely spread due to the frequent occurence of L0 source datasets with > 1 core observation variable.

This function collects specified columns from \code{L0_flat}, converts into long (attribute-value) form with \code{variable_name} names and values to the resulting table's "variable_name" and "value" columns, respectively. Regular expression matching joins \code{unit} to any associated \code{variable_name} and is listed in the resulting table's "unit" column.
}
\examples{
flat <- ants_L0_flat

location_ancillary <- create_location_ancillary(
  L0_flat = flat,
  location_id = "location_id",
  variable_name = "treatment")

location_ancillary

}
