% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observation.R
\name{create_observation}
\alias{create_observation}
\title{Create the observation table}
\usage{
create_observation(
  L0_wide,
  observation_id = "observation_id",
  event_id = NULL,
  package_id = "package_id",
  location_id = "location_id",
  datetime = "datetime",
  taxon_id = "taxon_id",
  variable_name,
  unit = NULL
)
}
\arguments{
\item{L0_wide}{(data.frame) The fully joined L0 dataset, in wide format.}

\item{observation_id}{(character) Column in \code{L0_wide} containing the the identifier assigned to each unique observation.}

\item{event_id}{(character; optional) Column in \code{L0_wide} containing the identifier assigned to each unique sampling event.}

\item{package_id}{(character) Column in \code{L0_wide} containing the identifier this dataset will have once it's published in a data repository.}

\item{location_id}{(character) Column in \code{L0_wide} containing the identifier assigned to each unique location at the observation level.}

\item{datetime}{(character) Column in \code{L0_wide} containing the date, and if applicable time, of the observation following the ISO-8601 standard format (i.e. YYYY-MM-DD hh:mm:ss).}

\item{taxon_id}{(character) Column in \code{L0_wide} containing the identifier assigned to each unique organism at the observation level.}

\item{variable_name}{(character) Columns in \code{L0_wide} containing the variables measured.}

\item{unit}{(character; optional) Columns in \code{L0_wide} containing the units of each \code{variable_name} following the column naming convention <unit>_<variable_name> (e.g. "unit_abundance").}
}
\value{
(data.frame) The location_ancillary table of ecocomDP.
}
\description{
Create the observation table
}
\details{
This function collects specified columns from \code{L0_wide}, converts into long (attribute-value) form with \code{variable_name} names and values to the resulting table's "variable_name" and "value" columns, respectively. Regular expression matching joins \code{unit} to any associated \code{variable_name} and is listed in the resulting table's "unit" column.
}
\examples{

}
