% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dwca.R
\name{create_tables_dwca_event_core}
\alias{create_tables_dwca_event_core}
\title{Creates DwC-A tables in event core}
\usage{
create_tables_dwca_event_core(
  dt_obs,
  dt_obs_ancil,
  dt_loc_ancil,
  dt_loc,
  dt_tax,
  parent.package.id,
  child.package.id
)
}
\arguments{
\item{dt_obs}{}

\item{dt_obs_ancil}{}

\item{dt_loc_ancil}{}

\item{dt_loc}{}

\item{dt_tax}{}

\item{parent.package.id}{(character) ID of an ecocomDP data package. Only 
EDI Data Repository package IDs are currently supported.}

\item{child.package.id}{}
}
\value{
three tables, event, occurrence, measurementOrFact
}
\description{
Creates DwC-A tables in event core
}
\examples{

}
