% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarwinCore_functions.R
\name{create_tables_dwca_event_core}
\alias{create_tables_dwca_event_core}
\title{Creates DwC-A tables in event core}
\usage{
create_tables_dwca_event_core(
  dwca_occurrence_core_config,
  dwca_occurrence_core_mapping,
  dt_obs,
  dt_obs_ancil,
  dt_loc_ancil,
  dt_loc,
  dt_tax
)
}
\arguments{
\item{dwca_occurrence_core_config}{}

\item{dwca_occurrence_core_mapping}{}

\item{dt_obs}{}

\item{dt_obs_ancil}{}

\item{dt_loc_ancil}{}

\item{dt_loc}{}

\item{dt_tax}{}
}
\value{
three tables, event, occurrence, measurementOrFact
}
\description{
Creates DwC-A tables in event core
}
\examples{
(dwca_event_table, dwca_occurrence_table, dwca_measurementOrFact_table) <- create_table_dwca_event_core(
 a bunch of inputs here.)    
}
