% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxon.R
\name{create_taxon}
\alias{create_taxon}
\title{Create the taxon table}
\usage{
create_taxon(
  L0_wide,
  taxon_id = "taxon_id",
  taxon_rank = "taxon_rank",
  taxon_name = "taxon_name",
  authority_system = "authority_system",
  authority_taxon_id = "authority_taxon_id"
)
}
\arguments{
\item{L0_wide}{(data.frame) The fully joined source L0 dataset, in wide format.}

\item{taxon_id}{(character) Column in \code{L0_wide} containing the identifier assigned to each unique organism at the observation level.}

\item{taxon_rank}{(character) The optional column in \code{L0_wide} containing the taxonomic rank of the organism in \code{taxon_name}.}

\item{taxon_name}{(character) Column in \code{L0_wide} containing the taxonomic name of the organism.}

\item{authority_system}{(character) An optional column in \code{L0_wide} containing the name of the authority system \code{authority_taxon_id} is from (e.g. "ITIS").}

\item{authority_taxon_id}{(character) An optional column in \code{L0_wide} containing the identifier corresponding to \code{taxon_name} in the \code{authority_system}.}
}
\value{
(data.frame) The taxon table of ecocomDP.
}
\description{
Create the taxon table
}
\details{
This function collects specified columns from \code{L0_wide} and returns distinct rows. Default names of optional columns are ignored if they can't be found in \code{L0_wide} (i.e. no need to set as NULL).

If taxon authority_system contains values of .... then the corresponding taxa will have their whole hierarchy expanded within the EML metadata returned by create_eml().
}
\examples{

}
