% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variable_mapping.R
\name{create_variable_mapping}
\alias{create_variable_mapping}
\title{Create the variable_mapping table}
\usage{
create_variable_mapping(
  observation,
  observation_ancillary = NULL,
  location_ancillary = NULL,
  taxon_ancillary = NULL
)
}
\arguments{
\item{observation}{(data.frame) The observation table.}

\item{observation_ancillary}{(data.frame) The optional observation_ancillary table.}

\item{location_ancillary}{(data.frame) The optional location_ancillary table.}

\item{taxon_ancillary}{(data.frame) The optional taxon_ancillary table.}
}
\value{
(data.frame) The variable_mapping table of ecocomDP.
}
\description{
Create the variable_mapping table
}
\details{
This function collects specified data tables, extracts unique variable_name values from each, converts into long (attribute-value) form with the table name and variable_name values to the resulting table's "table_name" and "variable_name" columns, respectively. The resulting table's "mapped_system", "mapped_id", and "mapped_label" are filled with \code{NA} and are to be manually filled.
}
\examples{

}
