% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables}
\alias{define_variables}
\title{Define variables for ecocomDP tables}
\usage{
define_variables(path, delimiter)

    Run this function to define variables prior to creating EML with 
    \code{make_eml}.
}
\arguments{
\item{path}{A path to the dataset working directory containing ecocomDP tables.}

\item{delimiter}{The field delimiter of ecocomDP tables. Can be comma or tab delimited 
(i.e. "," or "\t")}
}
\value{
A tab delimited UTF-8 file in the ecocomDP working directory titled 
    \emph{variables_tablename.txt} containing unique values 
    of attributes of class "categorical" which is translated and written 
    to EML with \code{make_eml}.
}
\description{
Variables of ecocomDP are stored as categorical variables in long format.
    Use this function to identify unique variables and assign definitions 
    and any associated units.
}
\details{
This function overwrites any 
    \emph{variables_tablename.txt} files you have created in the ecocomDP 
    working directory. To prevent overwriting of these files, temporarily 
    move them out of the working directory.
    
    This function identifies unique variables listed in the ecocomDP tables
    and writes them to a table (e.g. \emph{taxon_ancillary_variables}) for 
    you to define. Do not define categorical variables with empty
    field contents. Delete these rows from this file.
    
    When defining categorical variables with unit values, refer to the 
    standard unit dictionary "name" column. The unit dictionary will be 
    automatically opened when running this function. Enter the unit name in 
    the unit column of the categorical variables table. Note these values 
    are case sensitive. If you cannot find a unit in the dictionary, enter 
    the unit in the tab delimited UTF-8 formatted file 
    \emph{custom_units.txt} that has been automatically imported into your
    working directory. Valid custom units must be convertible to SI Units 
    (i.e. International System of Units). If it cannot be converted to SI 
    then list it units column as "dimensionless". To create a custom unit 
    define the:
    \itemize{
        \item \strong{id} This is equivalent to the unit name. Reference 
        the standard unit dictionary formatting
        \item \strong{unitType} The type of unit being defined.
        Reference the dictionary for examples.
        \item \strong{parentSI} The SI equivalent of the id you have entered.
        \item \strong{multiplierToSI} This is the multiplier to convert 
        from your custom unit to the SI unit equivalent.
        \item \strong{description} A description of the custom unit. 
        Reference the dictionary for examples.
        }
}
