% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables}
\alias{define_variables}
\title{Define categorical variables}
\usage{
define_catvars(path)

    Run this function to set variable definitions prior to running make_eml.
}
\arguments{
\item{path}{A path to the dataset working directory containing ecocomDP tables.}

\item{delimiter}{The field delimiter of ecocomDP tables. Can be comma or tab delimited 
(i.e. "," or "\t")}
}
\value{
A tab delimited UTF-8 file in the ecocomDP working directory titled 
    \emph{tablename_catvars.txt} containing unique values 
    of attributes of class "categorical" which is translated and written 
    to EML with \code{make_eml}.
}
\description{
Identify and define categorical variables of ecocomDP tables.
}
\details{
This function overwrites any 
    \emph{tablename_catvars.txt} files you have created in 
    the ecocomDP working directory. To prevent overwriting of these files, 
    temporarily move them out of the working directory.
    
    This function identifies "categorical" class attributes from the file 
    \emph{attributes_tablename.txt} and extracts unique 
    values of these attributes to the table for you to define. Do not define 
    categorical variables with empty field contents. Delete these rows from 
    this file.
    
    When defining categorical variables with unit values, refer to the 
    standard unit dictionary "name" column. Enter the unit name in the 
    definition column of the categorical variables table. Note these values 
    are case sensitive.
}
\seealso{
\code{\link{make_eml}} to make an EML file and write to the working
    directory.
}
