% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{delete_from_queue}
\alias{delete_from_queue}
\title{Delete item from the ecocom-listener's queue}
\usage{
delete_from_queue(index)
}
\arguments{
\item{index}{(integer) Index of item in queue to remove}
}
\value{

}
\description{
Delete item from the ecocom-listener's queue
}
