% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{detect_data_type}
\alias{detect_data_type}
\title{determine data type}
\usage{
detect_data_type(data)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset (list), or a list of ecocomDP tables, a single flat ecocomDP table, or a single flat table created by joinng and spreading ecocomDP tables.}
}
\value{
(character) A character string indicating data type. One of: 'list_of_tables', 'datset_old', 
'dataset', or 'flat_table'
}
\description{
determine data type
}
\details{
Internal function for detecting datatype. Not exported
}
