% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_source_conversion_script.R
\name{download_and_source_conversion_script}
\alias{download_and_source_conversion_script}
\title{Download and source L0 to L1 conversion script}
\usage{
download_and_source_conversion_script(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML of newest L1 child. The L0-to-L1 conversion script listed in \code{eml} will be used to create the child of the newest L0.}

\item{path}{(character) Directory to which create_ecocomDP.R will be written. Should be the same directory tables and metadata will be written to.}
}
\description{
Download and source L0 to L1 conversion script
}
\details{
The script is downloaded \code{path} and then parsed to identify R libraries used by the script which are installed if not already.
}
