% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{get_from_queue}
\alias{get_from_queue}
\title{Get next item in the ecocom-listener's queue}
\usage{
get_from_queue()
}
\value{
\item{index}{(integer) Index of item in queue. Is later used for removing the item from the queue.}
\item{id}{(character) Data package identifier}
\item{config.environment}{(character) Location of the \code{id} within a repository system. This variable is written to the Global Environment for use in \code{routine_handler()}.}
}
\description{
Get next item in the ecocom-listener's queue
}
