% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_L0_id.R
\name{get_previous_L0_id_and_L1_conversion_script}
\alias{get_previous_L0_id_and_L1_conversion_script}
\title{Get previous L0 data package identifier and URL link to conversion script}
\usage{
get_previous_L0_id_and_L1_conversion_script(
  package.id,
  repository = "EDI",
  environment = "production"
)
}
\arguments{
\item{package.id}{(character) Newest L0 data package identifier}

\item{repository}{(character) Data repository in which \code{package.id} resides and associated with \code{environment}. Currently supported repositories are: "EDI" (Environmental Data Initiative). Requests for support of other repositories can be made via \href{https://github.com/EDIorg/ecocomDP}{ecocomDP GitHub} issues. Default is "EDI".}

\item{evironment}{(character) Repository environment in which \code{package.id} exists. Some repositories have development, staging, and production environments which are distinct from one another. This argument allows reading of EML from different environments. Default is "production".}
}
\value{
(list) A list containing:
\item{id}{Identifier of previous L0 data package}
\item{script}{Download URL of the script used to convert the L0 to L1}
Otherwise NULL
}
\description{
Get the most recent L0 data package identifier containing a recognizable L0 to L1 conversion script and get the URL to download the script as well.
}
\details{
For the most recent L0 data package, which may not have a corresponding L1, follow all provenance traces to data packages containing the "ecocomDP" keyword. If more than one, select the most recent L0 identifier, read the corresponding EML, look for a valid L0 to L1 conversion script (convert_ecocomDP.R), and break if found. Otherwise try this process again for the next oldest L0 data package and so on.
}
