% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_child.R
\name{has_child}
\alias{has_child}
\title{Does a data package have a ecocomDP or DwC-A child?}
\usage{
has_child(keyword, package.id)
}
\arguments{
\item{keyword}{(character) Specifies the child type. Can be: "ecocomDP" or "Darwin Core Archive (DwC-A) Event Core"}

\item{package.id}{(character) Data package identifier}
}
\value{
(named logical) Named after the NEWEST child data package identifier
descending from \code{package.id}. It's possible to have multiple children
REVISIONS originating from a \code{package.id}. This may occur when there 
are issues with the first child of \code{package.id} that are fixed by a 
follow on revision.
}
\description{
Does a data package have a ecocomDP or DwC-A child?
}
