% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_templates.R
\name{import_templates}
\alias{import_templates}
\title{Import ecocomDP template files}
\usage{
import_templates(data.path)
}
\arguments{
\item{data.path}{A path to the dataset working directory.}
}
\value{
\emph{additional_contact.txt} A tab delimited UTF-8 text file in which 
    to provide contact information of the person that created the ecocomDP
    from a parent data package. This additional contact is added to the EML
    of the ecocomDP along with the contacts of the parent data package.
    
    \emph{cusom_units.txt} A UTF-8 text file containing fields to fill out
    for each custom unit that is found within the dataset.
}
\description{
Imports template files utilized in the ecocomDP creation process.
}
\details{
This function should be run at the beginning of the ecocomDP creation 
    process. Make a working directory for your dataset then run this function 
    to import template files. If template files already exist in the working 
    directory, new templates will not be imported.
}
