% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_templates.R
\name{import_templates}
\alias{import_templates}
\title{Import templates for ecocomDP}
\usage{
import_templates(path)
}
\arguments{
\item{path}{A path to the dataset working directory.}
}
\value{
\emph{map.txt} A tab delimited UTF-8 text file to map level-0 to level-1
    data. Open this file with a spreadsheet editor. and reference 
    instructions for help.
    
    \emph{map_instructions.txt} A UTF-8 text file containing instructions on
    how to complete mapping of level-0 to level-1.
}
\description{
Create a working directory for your dataset then run this function to 
    import templates. Use these templates to map level-0 data (raw) to 
    level-1 (ecocomDP), as well as provide metadata for your dataset.
}
\details{
If template files already exist in the working directory, new templates 
    will not be imported.
}
