% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_to_L2_DwCA.R
\name{L1_to_L2_DwCA}
\alias{L1_to_L2_DwCA}
\title{Creates Darwin Core Archive package from an ecocomDP formatted level-1}
\usage{
L1_to_L2_DwCA(path, core.name, parent.package.id, child.package.id)
}
\arguments{
\item{path}{(character) Path to the where the DwC-A data objects and EML will be 
written.}

\item{core.name}{(character) The Darwin Core central table of the package. Can be: 
"occurence" (occurrence core) or "event" (event core).}

\item{parent.package.id}{(character) ID of an ecocomDP data package. Only 
EDI Data Repository package IDs are currently supported.}

\item{child.package.id}{(character) ID of DWcA occurrence data package being created.}
}
\value{
Data entities in either the occurrence or event core formats with a
    corresponding EML metadata record and a meta.xml file.
}
\description{
Creates Darwin Core Archive package from an ecocomDP formatted level-1
}
\details{
TODO: Add details
}
\examples{

}
