% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{make_eml}
\alias{make_eml}
\title{Make EML for a set of ecocomDP tables}
\usage{
make_eml(data.path, code.path = data.path, code.files, 
    parent.package.id, sep, cat.vars = NULL, intellectual.rights = NULL, 
    child.package.id, user.id, affiliation, access.url = NULL, 
    additional.contact)
}
\arguments{
\item{data.path}{(character) Path to the directory containing ecocomDP tables.}

\item{code.path}{(character) Path to the directory containing scripts that reformat the 
parent data package to the child data package.}

\item{code.files}{(character) Path to the directory containing scripts that reformat the 
parent data package to the child data package.}

\item{parent.package.id}{(character) Parent data package ID (e.g. "knb-lter-hfr.118.28") in the 
EDI Data Repository.}

\item{sep}{(character) Field delimiter used in the ecocomDP tables. Valid options 
are "," or "\\t".}

\item{cat.vars}{(data frame) Categorical variables with associated definitions and 
units. Create a cat.vars data frame with the `define_variables` 
function. Variables that can't be defined by the parent package EML 
should be manually supplied to the cat.vars data frame.}

\item{intellectual.rights}{(character) The intellectual rights license to be used with the 
ecocomDP. If this argument is NULL, then the parent data package 
intellectual rights will be used. Valid options are "CCO" 
(https://creativecommons.org/publicdomain/zero/1.0/legalcode) and
"CCBY" (https://creativecommons.org/licenses/by/4.0/legalcode).}

\item{child.package.id}{(character) Child data package ID (e.g. "edi.53.1") to be uploaded to 
EDI Data Repository.}

\item{user.id}{(character) User ID of person uploading the ecocomDP data package to the
EDI Data Repository.}

\item{affiliation}{(character) Affiliation of the user.id. Valid options are "EDI" or 
"LTER".}

\item{access.url}{(character) The base URL access point to ecocomDP tables and scripts 
that can be downloaded by the EDI Data Repository (e.g.
https://lter.limnology.wisc.edu/sites/default/files/data). Alternatively
tables and scripts can be uploaded through the EDI Data Repository 
Portal.}

\item{additional.contact}{(data frame) Contact information for creator of the script that converts
the parent data package to the ecocomDP. The data frame must have these
columns:
\itemize{
    \item{givenName}
    \item{surName}
    \item{organizationName}
    \item{electronicMailAddress}
}}
}
\value{
An EML metadata record for the ecocomDP tables.
}
\description{
Makes EML for a set of ecocomDP tables.
}
\details{
This function creates valid EML for an Ecological Community Data Pattern 
    (ecocomDP) using a combination of elements from the parent data package, 
    boiler plate metadata specific to the ecocomDP, and user supplied 
    defined in the conversion script.
    
    Run \code{make_eml} after you have:
    \itemize{
        \item Created ecocomDP tables through a fully scripted process.
        \item Validated your ecocomDP with `validate_ecocomDP`.
        \item Defined variables contained in your ecocomDP using 
        `define_variables`.`
    }

    This function uses elements of the parent data package EML to create EML
    for the ecocomDP. Since ecocomDP is a fixed format, this function calls
    on boiler plate table attributes of the R package inst/ directory and 
    imports the metadata of the parent data package. Some new information is
    added to the parent data package EML as specified below.
    \itemize{
        \item \strong{<access>} Appends the ecocomDP creator to the list of 
        principals granted read and write access to the ecocomDP data 
        package.
        \item \strong{<pubDate>} Adds a publication date corresponding to 
        when the ecocomDP EML was created.
        \item \strong{<keywordSet} Appends two ecocomDP specific keyword 
        sets to the keyword sets listed in the parent data package EML.
        \item \strong{<intellectualRights>} Keeps intact the intellectual
        rights license of the parent data package, or replaces it with one
        of the user specified options above.
        \item \strong{<taxonomicCoverage>} Updates the taxonomic coverage 
        element with data supplied in the taxon table of the ecocomDP.
        \item \strong{<contact>} Appends contact information of the ecocomDP
        creator to the contacts listed in the parent data package.
        \item \strong{<methodStep>} Adds provenance metadata as a method 
        step, describing the parent data package from which the ecocomDP 
        was created.
        \item \strong{<dataTables>} Replaces the parent data package data
        tables with the newly created ecocomDP tables. NOTE: All other data
        entities are removed from the ecocomDP EML.
        \item \strong{<otherEntity>} Describes formatting scripts under the 
        other entity element.
    }
}
