% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{make_eml}
\alias{make_eml}
\title{Make EML}
\usage{
make_eml(path, parent.package.id, child.package.id, delimiter, user.id,
    author.system, intellectual.rights)
}
\arguments{
\item{path}{A path to the dataset working directory containing the validated 
ecocomDP tables.}

\item{parent.package.id}{Identifier of parent data package in LTER or EDI data repository (e.g. 
knb-lter-hfr.118.28).}

\item{child.package.id}{Identifier of child data package in LTER or EDI data repository (e.g. 
edi.53.1). If you don't have a child.package.id then use the default
"edi.100.1".}

\item{delimiter}{Field delimiter of input dataset. Can be comma (i.e. ",") or tab 
(i.e. "\t").}

\item{user.id}{LTER or EDI ID of person publishing the ecocomDP data package. If you 
don't have a user ID then use the default "EDI".}

\item{author.system}{The author system the user.id is associated with. If you don't have an 
author system specification use the default value "edi"}

\item{intellectual.rights}{The intellectual rights license to be used with the child ecocomDP data
package. Valid arguments are "CCO" (https://creativecommons.org/publicdomain/zero/1.0/legalcode),
"CCBY" (https://creativecommons.org/licenses/by/4.0/legalcode), and no 
argument. No argument input indicates the license wil remain the same
as the parent data package.}

\item{access.url}{This is the base URL that PASTA uses to access the ecocomDP tables.
For example, the base URL of a table that has a URL of 
https://lter.limnology.wisc.edu/sites/default/files/data/gleon_chloride/gleon_chloride_concentrations.csv
would have a base URL of 
https://lter.limnology.wisc.edu/sites/default/files/data/gleon_chloride.}
}
\value{
An EML metadata file written to the dataset working directory titled 
    \emph{packageID.xml}.
}
\description{
Make EML for an Ecological Community Data Pattern (ecocomDP) using 
    elements from the parent data package. This function does not yet 
    support creation of data packages that are not already published in the
    LTER (Long-term Ecological Research) or EDI (Environmental Data 
    Initiative) data repositories.
}
\details{
This script modifies elements of the parent data package EML. 
    Specifically this script:
    \itemize{
        \item \strong{Modifies access} Modifies access based on user supplied
        arguments.
        \item \strong{Publicaton date} Changes the publication date to when 
        the ecocomDP was created.
        \item \strong{Keywords} Adds two keyword sets which include the 
        key term "ecocomDP" and a set of keywords relevant to community data
        from the LTER controlled vocabulary.
        \item \strong{Intellectual rights} Updates the intellectual rights
        according to user specified arguments.
        \item \strong{Taxon coverage} This has not yet been implemented.
        \item \strong{Contact} Adds the ecocomDP creator as a contact.
        \item \strong{Provenance} Adds the provenance snippet from the 
        parent data package.
        \item \strong{Data tables} Adds data table elements for ecocomDP tables
        present.
        \item \strong{Formatting scripts} Adds formatting scripts used to 
        create the ecocomDP as otherEntitie.
    }
}
