% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{make_eml}
\alias{make_eml}
\title{Make EML metadata for ecocomDP tables}
\usage{
make_eml(
  path,
  parent.package.id,
  child.package.id,
  script,
  script.description,
  is.about = NULL,
  cat.vars,
  contact,
  user.id,
  user.domain,
  basis.of.record = NULL,
  url = NULL
)
}
\arguments{
\item{path}{(character) Path to the directory containing ecocomDP data tables, conversion scripts, and where EML metadata will be written.}

\item{parent.package.id}{(character) Parent data package ID (e.g. "knb-lter-hfr.118.28"). Only 
EDI Data Repository package IDs are currently supported.}

\item{child.package.id}{(character) Child data package ID (e.g. "edi.53.1") to be uploaded to 
EDI Data Repository.}

\item{script}{(character) Names of scripts used to convert and repackage the source 
data into the ecocomDP.}

\item{script.description}{(character) A description for each object listed under \code{script}.}

\item{is.about}{(named character) Dataset level annotations describing what this dataset "is about" (L0 keywords converted to object label + URI pairs) describing what the L1 dataset is about. (e.g. \code{is.about = c('level of ecological disturbance' = "http://purl.dataone.org/odo/ECSO_00002588",' type of ecological disturbance' = "http://purl.dataone.org/odo/ECSO_00002589")}).}

\item{cat.vars}{(data frame) Categorical variables with associated definitions and 
units. Create a cat.vars data frame with the 
\code{ecocomDP::define_variables()} function. Variables that can't be 
defined by the parent package EML should be manually supplied to the 
cat.vars data frame in one of the scripts specified in the 
\code{script} argument.}

\item{contact}{(data frame) Contact information of this ecocomDP creator.
The data frame must have these columns:
\itemize{
    \item{givenName}
    \item{surName}
    \item{organizationName}
    \item{electronicMailAddress}
}}

\item{user.id}{(character) ID(s) of user account(s) associated with the data repository
in which this ecocomDP data package will be archived. Only EDI Data 
Repository IDs are currently supported.}

\item{user.domain}{(character) Domain of the \code{user.id}(s). Only "EDI" and "LTER" are 
currently supported. If more than one, then supply as a vector of 
character strings in the same order as the corresponding \code{user.id}.}

\item{basis.of.record}{(character) To create a Darwin Core record from this ecocomDP using
\code{L1_to_L2_DwCA()} then the Darwin Core property 
\href{basisOfRecord}{https://dwc.tdwg.org/terms/#dwc:basisOfRecord}
requires definition as 
\href{HumanObservation}{http://rs.tdwg.org/dwc/terms/HumanObservation} or 
\href{MachineObservation}{http://rs.tdwg.org/dwc/terms/MachineObservation}.}

\item{url}{(character) URL to the publicly accessible directory containing ecocomDP data tables, conversion scripts, and EML metadata. This argument supports direct download of the data entities by a data repository and is used within the scope of the ecocomDP project for automated revision and upload of ecocomDP data packages and derived products.}
}
\value{
An EML metadata record for the ecocomDP dataset defined by the arguments
    to this function.
}
\description{
Run this function AFTER you have validated your ecocomDP tables with 
\code{ecocomDP::validate_ecocomDP()}.
}
\details{
This function creates an EML record for an Ecological Community Data 
    Pattern (ecocomDP) combining metadata from the parent data package and
    boiler-plate metadata describing the ecocomDP tables. Changes to the 
    parent EML include:
    \itemize{
        \item \strong{<access>} Adds the \code{user.id} to the list of 
        principals granted read and write access to the ecocomDP data 
        package this EML describes.
        \item \strong{<title>} Adds a note that this is a derived data 
        package in the ecocomDP format.
        \item \strong{<pubDate>} Adds the date this EML was created.
        \item \strong{<abstract>} Adds a note that this is a derived data 
        package in the ecocomDP format.
        \item \strong{<keywordSet} Adds the "ecocomDP" keyword to enable
        search and discovery of all ecocomDP data packages in the EDI Data
        Repository, and 7 terms from the LTER Controlled vocabulary:
        "communities", "community composition", "community dynamics", 
        "community patterns", "species composition", "species diversity",
        and "species richness". Darwin Core Terms listed under \code{
        basis.of.record} are listed and used by \code{L1_to_L2_DwCA()} 
        to create a Darwin Core Archive of this ecocomDP data package.
        \item \strong{<intellectualRights>} Keeps intact the intellectual
        rights license of the parent data package, or replaces it with
        "CCO" (https://creativecommons.org/publicdomain/zero/1.0/legalcode).
        \item \strong{<taxonomicCoverage>} Updates the taxonomic coverage 
        element with data supplied in the taxon table of the ecocomDP.
        \item \strong{<contact>} Adds contact information of the ecocomDP
        creator to the list of contacts in the parent data package EML.
        \item \strong{<methodStep>} Adds a note that this data package was
        created by the scripts listed under the \code{script} argument,
        and adds provenance metadata noting that this is a derived data 
        and describing where the parent data package can be accessed.
        \item \strong{<dataTables>} Replaces the parent data package data
        tables metadata with boiler-plate metadata for the ecocomDP tables.
        \item \strong{<otherEntity>} Describes scripts listed in the 
        \code{script} and \code{script.description} arguments. Any other
        entities listed in the parent EML are removed.
    }
}
\examples{
\dontrun{

# Set path to ecocomDP tables and scripts
path <- "/Users/csmith/Desktop/ecocomDP"

# Validate ecocomDP tables
ecocomDP::validate_ecocomDP(path)

# Use parent EML to define variables of the ecocomDP and manually add variable 
# names and definitions created in the ecocomDP re-formatting process.
catvars <- ecocomDP::define_variables(path, "knb-lter-sev.29.12")
catvars$unit[catvars$code == "Count"] <- "number"
catvars$definition[catvars$code == "Comments"] <- "A special statement related to an observation"

# Create contact information for the ecocomDP creator
additional_contact <- data.frame(
  givenName = 'Colin',
  surName = 'Smith',
  organizationName = 'Environmental Data Initiative',
  electronicMailAddress = 'csmith@wisc.edu',
  stringsAsFactors = FALSE
)

# Create EML
ecocomDP::make_eml(
  path = path,
  parent.package.id = "knb-lter-sev.29.12",
  child.package.id = "edi.101.5",
  script = c(
    "convert_sev29_to_ecocomDP.R", 
    "package_edi_333.R"
  ),
  script.description = c(
    "R script for creating the ecocomDP tables.",
    "R script for creating the ecocomDP EML"
  ),
  cat.vars = catvars,
  contact = additional_contact,
  user.id = 'csmith',
  user.domain = 'LTER',
)
}

}
