% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{make_eml}
\alias{make_eml}
\title{Make EML (WARNING: this function is not yet functional)}
\usage{
make_eml(path, parent.package.id, child.package.id)
}
\arguments{
\item{path}{A path to the dataset working directory containing the validated 
ecocomDP tables.}

\item{parent.package.id}{Identifier of parent data package (e.g. knb-lter-hfr.118.28).}

\item{child.package.id}{Identifier of child data package (e.g. edi.53.1).}
}
\value{
EML for the ecocomDP tables to the input path.
}
\description{
Make EML for an Ecological Community Data Pattern (ecocomDP) using 
    elements from the parent data package.
}
\details{
Details.
}
