% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml_neon.R
\name{make_eml_neon}
\alias{make_eml_neon}
\title{make_eml_neon}
\usage{
make_eml_neon(eml, x)
}
\arguments{
\item{eml}{(eml) An EML record for a NEON data product.}

\item{x}{(list of data frames) The ecocomDP tables for the NEON data product.
Each item of the list must be named after the ecocomDP tables (e.g.
observation, taxon, dataset_summary, etc.).}

\item{dp.id}{(character) NEON data product ID}
}
\value{
EML metadata for a NEON data product.
}
\description{
Make an EML metadata record for a NEON data product that has been 
    reformatted into the ecocomDP.
}
