% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location.R
\name{make_location}
\alias{make_location}
\title{Make location table}
\usage{
make_location(x, cols = c('site', 'nested_level_1', 'nested_level_2', 
    'etc.' ))
}
\arguments{
\item{x}{A data frame containing sampling location names (in columns) for each 
observation (rows).}

\item{cols}{A list of columns in x containing sampling location data. The list 
order should reflect the hierarchical arrangement of sampling locations.
E.g.: If 'site' is the coarsest level of observation, 'habitat' is 
nested within 'site', 'transect' is nested within 'habitat', and 
'quadrat' is nested within 'transect' then the 'cols' argument should
be supplied as:

cols = c('site', 'habitat','transect', 'quadrat')}
}
\value{
A data frame of the location table.
}
\description{
Make a data frame of the location table.
}
