% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxon.R
\name{make_taxon}
\alias{make_taxon}
\title{Make the taxon table}
\usage{
make_taxon(
  x = NULL,
  cols = NULL,
  taxa = NULL,
  name.type = "scientific",
  data.sources = NULL
)
}
\arguments{
\item{x}{(data frame) A data frame containing taxa names (in one or more 
columns).}

\item{cols}{(character) One or more columns in \code{x} that when combined form a 
unique taxa that may be resolvable to a taxonomic authority system. 
(e.g. \code{cols = c("genus", "species")}).}

\item{taxa}{(character) If not using \code{x} and \code{cols}, then a list of 
unique taxa can be input using this argument.}

\item{name.type}{(character) Taxonomic name type. Can be: 'scientific', 'common', or 
'both'.}

\item{data.sources}{(integer) An ordered numeric vector of ID's corresponding to taxonomic 
authorities to resolve taxa to (in the order of decreasing preference). 
Run \code{view_taxa_authorities()} to see supported 
authorities.}
}
\value{
A list containing:
\item{taxon}{A data frame of the taxon table}
\item{x}{The input \code{x} with an added taxon_id column linking unique
\code{cols} to the list of taxa in the taxon table.}
}
\description{
Make the taxon table
}
