% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_variable_mapping.R
\name{make_variable_mapping}
\alias{make_variable_mapping}
\title{Make variable_mapping table}
\usage{
make_variable_mapping(
  observation,
  observation_ancillary = NULL,
  location_ancillary = NULL,
  taxon_ancillary = NULL
)
}
\arguments{
\item{observation}{(data frame) The completed observation table.}

\item{observation_ancillary}{(data frame) The completed observation_ancillary table.}

\item{location_ancillary}{(data frame) The completed location_ancillary table.}

\item{taxon_ancillary}{(data frame) The completed taxon_ancillary table.}
}
\value{
A data frame of the variable_mapping with completed columns:
    \itemize{
        \item{variable_mapping_id}
        \item{table_name}
        \item{variable_name}
    }
    Other columns are filled with NA.
}
\description{
Make a data frame of the variable mapping table.
}
