% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetime}
\usage{
parse_datetime(vals, frmt)
}
\arguments{
\item{vals}{(character) Vector of datetimes}

\item{frmt}{(character) Datetime format string}
}
\value{
(POSIXct POSIXt) Datetimes parsed
}
\description{
Parse datetime
}
\details{
A wrapper to to \code{lubridate::ymd()}, \code{lubridate::ymd_h()}, \code{lubridate::ymd_hm()}, and \code{lubridate::ymd_hms()}.
}
