% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot alpha diversity (taxa richness) over time and space}
\usage{
plot_alpha_diversity(observation, id, alpha = 1)
}
\arguments{
\item{observation}{(data.frame) The observation table of ecocomDP}

\item{id}{(character) Dataset id. Used for plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. For when many data points overlap. Between 0 and 1, where 1 is 100\% opaque.}
}
\description{
Plot alpha diversity (taxa richness) over time and space
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
