% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot alpha diversity (taxa richness) over time and space}
\usage{
plot_alpha_diversity(observation, id, alpha = 1)
}
\arguments{
\item{observation}{(data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\description{
Plot alpha diversity (taxa richness) over time and space
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
