% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot alpha diversity (taxa richness) over time and space}
\usage{
plot_alpha_diversity(dataset, alpha = 1)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\description{
Plot alpha diversity (taxa richness) over time and space
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
