% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot community metrics}
\usage{
plot_data(dataset, type, path, alpha = 1)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}

\item{type}{(character) Type of plots to create. Options: "community"}

\item{path}{(character) Path to directory where plots will be written}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\value{
\item{alpha_diversity.pdf}{Alpha diversity (species richness) over time and space}
\item{sampling_effort.pdf}{Spatiotemporal sampling effort}
\item{sp_accumulation_over_space.pdf}{Species accumulation curve over space}
\item{sp_accumulation_over_time.pdf}{Species accumulation curves over time (site-specific and total)}
\item{sp_shared_among_sites.pdf}{Species shared among sites}
}
\description{
Plot community metrics
}
\details{
LIST METHODS OF CALCULATION FOR EACH PLOT
}
\note{
LIST DATA AGGREGATION REQUIREMENTS FOR THIS FUNC TO WORK.
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
