% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_sampling_times}
\alias{plot_sampling_times}
\title{Plot spatiotemporal sampling effort}
\usage{
plot_sampling_times(dataset, alpha = 1)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100\% opaque}
}
\description{
Plot spatiotemporal sampling effort
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
