% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_sites}
\alias{plot_sites}
\title{Plot sites on US map}
\usage{
plot_sites(data, id = NA_character_, alpha = 1, labels = TRUE)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, a flattened ecocomDP dataset, or any flat table containing the columns present in both the ecocomDP observation and location tables.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}

\item{labels}{(boolean) Argument to show labels of each US state. Default is TRUE.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot sites on US map
}
\examples{
\dontrun{
# plot map of sites
plot_sites(ants_L1)

plot_sites(flatten_data(ants_L1))


# download and plot NEON macroinvertebrate data
my_dataset <- read_data(
  id = "neon.ecocomdp.20120.001.001",
  site= c('COMO','LECO'), 
  startdate = "2017-06",
  enddate = "2019-09",
  check.size = FALSE)

plot_sites(my_dataset)
}

}
