% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_sites}
\alias{plot_sites}
\title{Plot sites on US map}
\usage{
plot_sites(
  dataset = NULL,
  flat_data = NULL,
  id = NA_character_,
  observation = NULL,
  location = NULL,
  alpha = 1,
  labels = TRUE
)
}
\arguments{
\item{dataset}{(list) An ecocomDP formatted dataset}

\item{flat_data}{(tbl_df, tbl, data.frame) A flat data.frame created from an \code{ecocomDP} formatted dataset using \code{ecocomDP::flatten_data}.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{observation}{(tbl_df, tbl, data.frame) The observation table.}

\item{location}{(tbl_df, tbl, data.frame) The location table.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}

\item{labels}{(boolean) Argument to show labels of each US state. Default is TRUE.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot sites on US map
}
\examples{
\dontrun{
plot_sites(
  dataset = ants_L1)

plot_sites(
  flat_data = flatten_dataset(ants_L1))

plot_sites(
  observation = ants_L1[[1]]$tables$observation,
  location = ants_L1[[1]]$tables$location)
}
}
