% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_sites}
\alias{plot_sites}
\title{Plot sites on US map}
\usage{
plot_sites(flat_table, id, alpha = 1, labels = TRUE)
}
\arguments{
\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}

\item{labels}{(boolean) Argument to show labels of each US state. Default is TRUE.}

\item{flat_tables}{(list of tbl_df, tbl, data.frame) A flat table, the result of flatten_data() applied to the list of ecocomDP tables.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot sites on US map
}
\examples{
flat_table <- flatten_data(ants_L1[[1]]$tables)
id <- names(ants_L1)

plot_sites(flat_table, id)

}
