% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_abundance}
\alias{plot_taxa_abundance}
\title{Plot abundances by event_id}
\usage{
plot_taxa_abundance(
  data = NULL,
  id = NA_character_,
  min_abundance = 0,
  trans = "identity",
  facet_var = NA_character_,
  color_var = NA_character_,
  facet_scales = "free",
  alpha = 1
)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, a flattened ecocomDP dataset, or any flat table containing the columns present in both the ecocomDP observation and taxon tables.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{min_abundance}{(numeric) Minimum abundance allowed for observations of taxa included in the plot.}

\item{trans}{(character, "identity" is default, "log1p" is x+1 transform) For continuous scales, the name of a transformation object or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time".}

\item{facet_var}{(character) Name of column to use for faceting.}

\item{color_var}{(character) Name of column to use for plot colors.}

\item{facet_scales}{(character) Should scales be free ("free", default value), fixed ("fixed"), or free in one dimension ("free_x", "free_y")?}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot abundances by event_id
}
\examples{
\dontrun{
# plot ecocomDP formatted dataset
plot_taxa_abundance(
  dataset = ants_L1)

# plot flattened ecocomDP dataset, log(x+1) transform abundances
plot_taxa_abundance(
  flat_data = flatten_dataset(ants_L1),
  trans = "log1p")

# facet by location color by taxon_rank, log 10 transformed
plot_taxa_abundance(
  dataset = ants_L1,
  facet_var = "location_id",
  color_var = "taxon_rank",
  trans = "log10")

# facet by location, only plot taxa of rank = "species"
plot_taxa_abundance(
  dataset = ants_L1,
  facet_var = "location_id",
  rank = "Species", min_occurrence = 5,
  trans = "log1p")

# color by location, only include taxa with > 10 occurrences
plot_taxa_abundance(
  dataset = ants_L1,
  color_var = "location_id",
  trans = "log10")

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_dataset() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_abundance(flat_data = .,
                      trans = "log1p")
}
}
