% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_taxa_accum_sites}
\alias{plot_taxa_accum_sites}
\title{Plot taxa accumulation curve over space}
\usage{
plot_taxa_accum_sites(dataset, alpha = 1)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}

\item{alpha}{(numeric) Alpha-transparency scale between 0 and 1, where 1 is 100% opaque}
}
\description{
Plot taxa accumulation curve over space
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
