% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_accum_sites}
\alias{plot_taxa_accum_sites}
\title{Plot taxa accumulation by site accumulation}
\usage{
plot_taxa_accum_sites(observation, id, alpha = 1)
}
\arguments{
\item{observation}{(tbl_df, tbl, data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot taxa accumulation by site accumulation
}
\examples{
observation <- ants_L1[[1]]$tables$observation
id <- names(ants_L1)

plot_taxa_accum_sites(observation, id)

}
