% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_occurrence_frequency}
\alias{plot_taxa_occurrence_frequency}
\title{Plot stacked taxa by site}
\usage{
plot_taxa_occurrence_frequency(
  data,
  id = NA_character_,
  min_occurrence = 0,
  facet_var = NA_character_,
  color_var = NA_character_,
  facet_scales = "free",
  alpha = 1
)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, a flattened ecocomDP dataset, or any flat table containing the columns present in both the ecocomDP observation and taxon tables.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{min_occurrence}{(numeric) Minimum number of occurrences allowed for taxa included in the plot.}

\item{facet_var}{(character) Name of column to use for faceting.}

\item{color_var}{(character) Name of column to use for plot colors.}

\item{facet_scales}{(character) Should scales be free ("free", default value), fixed ("fixed"), or free in one dimension ("free_x", "free_y")?}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device.
}
\description{
Plot stacked taxa by site
}
\examples{
\dontrun{
# plot ecocomDP formatted dataset
plot_taxa_occurrence_frequency(ants_L1)

# plot flattened ecocomDP dataset
plot_taxa_occurrence_frequency(flatten_data(ants_L1))

# facet by location color by taxon_rank
plot_taxa_occurrence_frequency(
  data = ants_L1,
  facet_var = "location_id",
  color_var = "taxon_rank")

# color by location, only include taxa with > 10 occurrences
plot_taxa_occurrence_frequency(
  data = ants_L1,
  color_var = "location_id",
  min_occurrence = 5)

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_occurrence_frequency()
}

}
