% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_occurrence_frequency}
\alias{plot_taxa_occurrence_frequency}
\title{Plot stacked taxa by site}
\usage{
plot_taxa_occurrence_frequency(
  dataset = NULL,
  flat_data = NULL,
  id = NA_character_,
  rank = NA_character_,
  cutoff = 0,
  facet_var = NA_character_,
  color_var = NA_character_,
  facet_scales = "free",
  alpha = 1
)
}
\arguments{
\item{dataset}{(list) An ecocomDP formatted dataset}

\item{flat_data}{(tbl_df, tbl, data.frame) A flat data.frame created from an \code{ecocomDP} formatted dataset using \code{ecocomDP::flatten_data}.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{rank}{(string) The rank of taxa to plot. Defaults to NA, which will inlcude all ranks.}

\item{cutoff}{(numeric) Minimum number of occurrences allowed for taxa included in the plot.}

\item{facet_var}{(character) Name of column to use for faceting.}

\item{color_var}{(character) Name of column to use for plot colors.}

\item{facet_scales}{(character) Should scales be free ("free", default value), fixed ("fixed"), or free in one dimension ("free_x", "free_y")?}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device.
}
\description{
Plot stacked taxa by site
}
\examples{
\dontrun{
# plot ecocomDP formatted dataset
plot_taxa_occurrence_frequency(
  dataset = ants_L1)

# plot flattened ecocomDP dataset
plot_taxa_occurrence_frequency(
  flat_data = flatten_dataset(ants_L1))

# facet by location color by taxon_rank
plot_taxa_occurrence_frequency(
  dataset = ants_L1,
  facet_var = "location_id",
  color_var = "taxon_rank")

# facet by location, only plot taxa of rank = "species"
plot_taxa_occurrence_frequency(
  dataset = ants_L1,
  facet_var = "location_id",
  rank = "Species", cutoff = 2)

# color by location, only include taxa with > 10 occurrences
plot_taxa_occurrence_frequency(
  dataset = ants_L1,
  color_var = "location_id",
  cutoff = 5)

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_dataset() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_occurrence_frequency(flat_data = .)
}
}
