% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_rank}
\alias{plot_taxa_rank}
\title{Plot taxa ranks}
\usage{
plot_taxa_rank(
  dataset = NULL,
  facet_var = NA_character_,
  taxon = NULL,
  observation = NULL,
  id = NA_character_,
  alpha = 1
)
}
\arguments{
\item{dataset}{(list) An ecocomDP formatted dataset}

\item{taxon}{(tbl_df, tbl, data.frame) The taxon table.}

\item{observation}{(tbl_df, tbl, data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot taxa ranks
}
\examples{
observation <- ants_L1[[1]]$tables$observation
taxon <- ants_L1[[1]]$tables$taxon
id <- names(ants_L1)

plot_taxa_rank(observation, taxon, id)

}
