% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_sample_time}
\alias{plot_taxa_sample_time}
\title{Plot dates and times samples were taken (DEPRECATED)}
\usage{
plot_taxa_sample_time(observation, id = NA_character_, alpha = 1)
}
\arguments{
\item{observation}{(tbl_df, tbl, data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}

\item{alpha}{(numeric) Alpha-transparency scale of data points. Useful when many data points overlap. Allowed values are between 0 and 1, where 1 is 100\% opaque. Default is 1.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
This function has been deprecated. Use \code{plot_sample_space_time()} instead.
}
\keyword{internal}
