% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_shared_sites}
\alias{plot_taxa_shared_sites}
\title{Plot number of unique taxa shared across sites}
\usage{
plot_taxa_shared_sites(data, id = NA_character_)
}
\arguments{
\item{data}{(list or tbl_df, tbl, data.frame) An ecocomDP formatted dataset, an ecocomDP observation table, or any table containing the columns present in an ecocomDP observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot number of unique taxa shared across sites
}
\examples{
\dontrun{
# plot ecocomDP formatted dataset
plot_taxa_shared_sites(ants_L1)

# plot flattened ecocomDP dataset
plot_taxa_shared_sites(flatten_data(ants_L1))

# plot ecocomDP observation table
plot_taxa_shared_sites(ants_L1[[1]]$tables$observation)

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_shared_sites()

# tidy syntax, filter data by site ID
ants_L1 \%>\% 
  flatten_data() \%>\% 
  dplyr::filter(
    as.numeric(location_id) > 4) \%>\%
  plot_taxa_shared_sites()
}

}
