% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_taxa_shared_sites}
\alias{plot_taxa_shared_sites}
\title{Plot taxa shared among sites}
\usage{
plot_taxa_shared_sites(dataset)
}
\arguments{
\item{dataset}{(list) Data object returned by \code{read_data()} (? list of named datapackage$tables)}
}
\description{
Plot taxa shared among sites
}
\examples{
\dontrun{
# one dataset
# - summarizing sometimes required
# more than one dataset
}

}
