% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_shared_sites}
\alias{plot_taxa_shared_sites}
\title{Plot number of unique taxa shared across sites}
\usage{
plot_taxa_shared_sites(observation, id)
}
\arguments{
\item{observation}{(data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}
}
\description{
Plot number of unique taxa shared across sites
}
\examples{
observation <- ants_L1[[1]]$tables$observation
id <- names(ants_L1)

plot_taxa_shared_sites(observation, id)

}
