% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa_shared_sites}
\alias{plot_taxa_shared_sites}
\title{Plot number of unique taxa shared across sites}
\usage{
plot_taxa_shared_sites(
  dataset = NULL,
  flat_data = NULL,
  observation = NULL,
  id = NA_character_
)
}
\arguments{
\item{dataset}{(list) An ecocomDP formatted dataset}

\item{flat_data}{(tbl_df, tbl, data.frame) A flat data.frame created from an \code{ecocomDP} formatted dataset using \code{ecocomDP::flatten_data}.}

\item{observation}{(tbl_df, tbl, data.frame) The observation table.}

\item{id}{(character) Identifier of dataset to be used in plot subtitles.}
}
\value{
(gg, ggplot) A gg, ggplot object if assigned to a variable, otherwise a plot to your active graphics device
}
\description{
Plot number of unique taxa shared across sites
}
\examples{
# plot ecocomDP formatted dataset
plot_taxa_shared_sites(
  dataset = ants_L1)

# plot flattened ecocomDP dataset
plot_taxa_shared_sites(
  flat_data = flatten_dataset(ants_L1))

# plot ecocomDP observation table
plot_taxa_shared_sites(
  observation = ants_L1[[1]]$tables$observation)

# tidy syntax, filter data by date
ants_L1 \%>\% 
  flatten_dataset() \%>\% 
  dplyr::filter(
    lubridate::as_date(datetime) > "2003-07-01") \%>\%
  plot_taxa_shared_sites(flat_data = .)

# tidy syntax, filter data by site ID
ants_L1 \%>\% 
  flatten_dataset() \%>\% 
  dplyr::filter(
    as.numeric(location_id) > 4) \%>\%
  plot_taxa_shared_sites(flat_data = .)

}
