% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read an ecocomDP dataset}
\usage{
read_data(id, path)
}
\arguments{
\item{id}{(character) Identifier of dataset to read. IDs are listed in the id 
column of search results output from \code{search_data()}.}

\item{path}{(character; optional) Path to the directory in which the aggregated 
ecocomDP will be written.}
}
\value{
(list) A named list, including: 
\item{metadata}{A list of information about the data.}
\item{tables}{A list of data.frames following the ecocomDP format.}
}
\description{
Read an ecocomDP dataset
}
\details{
Each data source has a corresponding reader. Determine the source from 
    the id and call the correct reader.
}
\examples{
d <- read_data("edi.193.3")
d <- read_data("DP1.20166.001")
}
