% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read ecocomDP data}
\usage{
read_data(
  id,
  path,
  site = "all",
  startdate = NA,
  enddate = NA,
  check.size = FALSE,
  nCores = 1,
  forceParallel = FALSE
)
}
\arguments{
\item{id}{(character) Identifier of dataset to read. Identifiers are listed in the 
"id" column of the \code{search_data()} output.  If reading multiple 
datasets, then \code{id} should be a character vector of identifiers or 
a named list containing additional arguments to filter on (for NEON data 
only; see below for arguments and examples).}

\item{path}{(character) Path to the directory in which the data will be written as 
an .rda object.}

\item{site}{(character; NEON data only) A character vector of site codes to filter 
data on. Sites are listed in the "sites" column of the 
\code{search_data()} output. Defaults to "all", meaning all sites.}

\item{startdate}{(character; NEON data only) Start date to filter on in the form YYYY-MM. 
Defaults to NA, meaning all available dates.}

\item{enddate}{(character; NEON data only) End date to filter on in the form YYYY-MM. 
Defaults to NA, meaning all available dates.}

\item{check.size}{(logical; NEON data only) Should the user approve the total file size 
before downloading? Defaults to FALSE.}

\item{nCores}{(integer; NEON data only) The number of cores to parallelize the 
stacking procedure. Defaults to 1.}

\item{forceParallel}{(logical; NEON data only) If the data volume to be processed does not 
meet minimum requirements to run in parallel, this overrides. Defaults 
to FALSE.}
}
\value{
(list) A named list, including: 
    \item{metadata}{A list of information about the data.}
    \item{tables}{A list of data.frames following the ecocomDP format 
    (\url{https://github.com/EDIorg/ecocomDP/blob/master/documentation/model/table_visualization.md})}
}
\description{
Read ecocomDP data
}
\details{
Validation checks are applied to each dataset ensuring they comply with 
    the ecocomDP. A warning is issued when validation fails (please report 
    these to \url{https://github.com/EDIorg/ecocomDP/issues}). Datasets that
    pass validation are returned.
    
    Column classes are coerced to those defined in the ecocomDP 
    specification.
    
    Reading multiple datasets raises the issue of referential integrity. 
    This is addressed by adding the \code{id} to each table's primary key
    (e.g. observation_id, event_id, etc.).
}
\examples{
# Read a dataset
d <- read_data("edi.193.3")
d <- read_data("DP1.20166.001")

# Read a NEON dataset with filters
d <- read_data(
  id = "DP1.20166.001", 
  site = c("MAYF", "PRIN"), 
  startdate = "2016-1", 
  enddate = "2018-11")

# Read multiple datasets from different sources
d <- read_data(c("edi.193.3", "DP1.20166.001"))

# Read multiple datasets from EDI and NEON with NEON filters
d <- read_data(
  id = list(
    edi.193.3 = NULL,
    DP1.20166.001 = list(
      site = c("MAYF", "PRIN"),
      startdate = "2016-1",
      enddate = "2018-11",
      check.size = FALSE)))

}
