% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{Read EML metadata from a data repository}
\usage{
read_eml(package.id, repository = "EDI", environment = "production")
}
\arguments{
\item{package.id}{(character) Data package identifier}

\item{repository}{(character) Data repository in which \code{package.id} resides and associated with \code{environment}. Currently supported repositories are: "EDI" (Environmental Data Initiative). Requests for support of other repositories can be made via \href{https://github.com/EDIorg/ecocomDP}{ecocomDP GitHub} issues. Default is "EDI".}

\item{evironment}{(character) Repository environment in which \code{package.id} exists. Some repositories have development, staging, and production environments which are distinct from one another. This argument allows reading of EML from different environments. Default is "production".}
}
\value{
(xml_document, xml_node) EML metadata
}
\description{
A wrapper function to repository specific read methods (the repository arg drives the logic).
}
\examples{

}
