% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{read_from_files}
\alias{read_from_files}
\title{Read ecocomDP from files}
\usage{
read_from_files(data.path)
}
\arguments{
\item{data.path}{(character) The path to the directory containing ecocomDP tables. 
Duplicate file names are not allowed.}
}
\value{
(list) A named list of data frames, each with the contents of the 
    corresponding file. Names follow ecocomDP specification.
}
\description{
Read ecocomDP from files
}
\note{
This function will deprecate \code{read_ecocomDP_table()}
}
\examples{
d <- read_from_files(system.file("/data", package = "ecocomDP"))

}
