% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_from_files}
\alias{read_from_files}
\title{Read ecocomDP from files}
\usage{
read_from_files(data.path)
}
\arguments{
\item{data.path}{(character) The path to the directory containing ecocomDP tables. 
Duplicate file names are not allowed.}
}
\value{
(list) A named list of \code{id}, each including: 
    \item{metadata}{A list of information about the data.}
    \item{tables}{A list of data.frames following the ecocomDP format 
    (\url{https://github.com/EDIorg/ecocomDP/blob/master/documentation/model/table_visualization.md})}
}
\description{
Read ecocomDP from files
}
\examples{
d <- read_from_files(system.file("/data", package = "ecocomDP"))

}
