% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_neon.R
\name{reformat_neon}
\alias{reformat_neon}
\title{reformat_neon}
\usage{
reformat_neon(dp.id)
}
\arguments{
\item{dp.id}{(character) NEON data product ID (e.g. 'DP1.20120.001').}

\item{neon.sites}{(character) A list of NEON site abbreviations from which to get
data for.}
}
\value{
Core ecocomDP tables for a user supplied NEON data product.
}
\description{
Reformat NEON data products into the ecocomDP.
}
