% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_handler.R
\name{routine_handler}
\alias{routine_handler}
\title{Event notification handler}
\usage{
routine_handler(package.id, environment, index)
}
\description{
Checks for unprocessed items in SQlite and pops the longest waiting item into \code{update_L1()} or \code{update_L2_dwca()}.
}
