% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save ecocomDP data}
\usage{
save_data(data, path, file.type = ".rds", file.name = NULL)
}
\arguments{
\item{data}{(list) Data as a list object created by \code{read_data()}. Name of object will become file name if \code{file.name} is not used}

\item{path}{(character) Path to the directory in which the data will be written.}

\item{file.type}{(character) Type of file to save the data to. Default is ".rds" but can also be ".csv"}

\item{file.name}{(character) Use this to set the file name if you'd like to be different than \code{data}.}
}
\value{
\item{.rda}{If \code{file.type} = ".rda", then an .rda representation 
    of \code{data} is returned.}
    \item{.csv}{If \code{file.type} = ".csv", then an set of .csv files are
    written to a sub-directory of \code{path} named after the data 
    package/product ID.}
}
\description{
Save ecocomDP data
}
\examples{
d <- read_data("edi.193.3")
#' save_data(d, tempdir(), ".rda")
save_data(d, tempdir(), ".csv")

}
