% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{Search ecocomDP data}
\usage{
search_data(
  text,
  taxa,
  num.taxa,
  years,
  sd.between.surveys,
  geographic.area,
  boolean.operator = "AND"
)
}
\arguments{
\item{text}{(character) Text to search in dataset titles, descriptions, and 
abstracts.}

\item{taxa}{(character) Taxanomic rank values to search for. The full taxonomic 
hierarchy of each taxa in the data is searchable. Scientific and 
common names are supported.}

\item{num.taxa}{(numeric) Minimum and maximum number of taxa in the data.}

\item{years}{(numeric) Minimum and maximum number of years sampled.}

\item{sd.between.surveys}{(numeric) Minimum and maximum standard deviation between surveys (in 
years).}

\item{geographic.area}{(numeric) Decimal degree North, East, South, and West coordinates within
which the data shoud originate.}

\item{boolean.operator}{(character) Boolean operator to use when searching \code{text} and 
\code{taxa}. Supported operators are "AND", "OR".}
}
\value{
(data.frame) Search results with these feilds:
    \itemize{
        \item source - Source from which the data originate. Currently 
        supported are "EDI" and "NEON".
        \item id - Identifier of dataset
        \item title - Title of dataset
        \item description - Description of dataset
        \item abstract - Abstract of dataset
        \item years - Number of years sampled
        \item sd_between_samples - Standard deviation between sampling 
        events in years.
        \item sites - Sites names or abbreviations (returned for NEON data
        sources but not EDI)
        \item url - URL to dataset
    }
}
\description{
Use this function to search across ecocomDP data made available by the 
Environmental Data Initiative (EDI) and the National Ecological Observatory
Network (NEON).
}
\note{
Searches across input arguments are combined with the "AND" boolean 
    operator.
}
\examples{
        
}
