% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{Search ecocomDP data}
\usage{
search_data(
  text,
  taxa,
  num.taxa,
  years,
  sd.between.surveys,
  geographic.area,
  boolean.operator = "AND"
)
}
\arguments{
\item{text}{(character) Text to search in dataset titles, descriptions, and 
abstracts.}

\item{taxa}{(character) Taxanomic rank values to search for. The full taxonomic 
hierarchy of each taxa in the data is searchable. Scientific and 
common names are supported.}

\item{num.taxa}{(numeric) Minimum and maximum number of taxa in the data.}

\item{years}{(numeric) Minimum and maximum number of years sampled.}

\item{sd.between.surveys}{(numeric) Minimum and maximum standard deviation between surveys (in 
years).}

\item{geographic.area}{(numeric) Decimal degree North, East, South, and West coordinates within
which the data should originate.}

\item{boolean.operator}{(character) Boolean operator to use when searching \code{text} and 
\code{taxa}. Supported operators are "AND", "OR".}
}
\value{
(data.frame) Search results with these feilds:
    \itemize{
        \item source - Source from which the data originate. Currently 
        supported are "EDI" and "NEON".
        \item id - Identifier of dataset
        \item title - Title of dataset
        \item description - Description of dataset
        \item abstract - Abstract of dataset
        \item years - Number of years sampled
        \item sd_between_samples - Standard deviation between sampling 
        events in years.
        \item sites - Sites names or abbreviations (returned for NEON data
        sources but not EDI)
        \item url - URL to dataset
    }
}
\description{
Use this function to search across ecocomDP data made available by the 
Environmental Data Initiative (EDI) and the National Ecological Observatory
Network (NEON).
}
\details{
Currently, to accomodate multiple L1 versions of NEON data products, search results for a NEON L0 will also list all the L1 versions available for the match. This method is based on the assumption that the summary data among L1 versions is the same, which may need to be addressed in the future. A list of L0 and corresponding L1 identifiers are listed in /inst/L1_versions.txt. Each L1 version is accompanied by qualifying text that's appended to the title, abstract, and descriptions for comprehension of the differences among L1 versions.
}
\note{
Searches across input arguments are combined with the "AND" boolean 
    operator.
}
\examples{
# Empty search returns all available data
r <- search_data()

# Text searches titles, descriptions, and abstracts
r <- search_data(text = "Lake")

# Taxa searches any taxonomic rank of an organism
r <- search_data(taxa = "Plantae")

# Number of taxa searches the count of unique taxa in a dataset
r <- search_data(num.taxa = c(0, 10))

# Years searches the number of years the dataset covers
r <- search_data(years = c(10, 20))

# Standard deviation between surveys searches for a sampling frequency
r <- search_data(sd.between.surveys = c(.25, 1))

# Geographic area searches where data were collected
r <- search_data(geographic.area = c(47.1, -86.7, 42.5, -92))

# Boolean operators AND (default) & OR combine text and taxa search terms
r <- search_data(text = c("Lake", "River"))
r <- search_data(text = c("Lake", "River"), boolean.operator = "OR")
r <- search_data(taxa = c("Plantae", "Animalia"))
r <- search_data(taxa = c("Plantae", "Animalia"), boolean.operator = "OR")

# Use any combination of search fields to find the data you're looking for
r <- search_data(
  text = c("Lake", "River"),
  taxa = c("Plantae", "Animalia"),
  num.taxa = c(0, 10),
  years = c(10, 100),
  sd.between.surveys = c(.01, 100),
  geographic.area = c(47.1, -86.7, 42.5, -92),
  boolean.operator = "OR")
        
}
