% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data_neon.R
\name{summarize_data_neon}
\alias{summarize_data_neon}
\title{Summarize supported NEON data products}
\usage{
summarize_data_neon(path)
}
\arguments{
\item{path}{(character) Path the file will be saved to

This is a helper function for project maintainers. It summarizes supported NEON data products for \code{search_data()}. This object essentially serves as an index of searchable terms.}
}
\value{
\item{summary_data_neon.rda}{Saved to /data}
}
\description{
Summarize supported NEON data products
}
\details{
Currently, only one L1 is derived form a NEON data product but there is already talk of supporting multiple L1 versions to meet user preferences. To do so, the content of summary_data_neon.rda shouldn't have to change but parsing to the searchable object used in \code{search_data()} may. See \code{search_data()} details for more.
}
\examples{
\dontrun{
summarize_data_neon(path)
}
        
}
