% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_L1.R
\name{update_L1}
\alias{update_L1}
\title{Update L1 from updated L0}
\usage{
update_L1(
  package.id.L0,
  repository = "EDI",
  evironment = "production",
  path,
  url,
  user.id,
  user.pass
)
}
\arguments{
\item{package.id.L0}{(character) Identifier of updated L0.}

\item{repository}{(character) Data repository in which \code{package.id.L0} resides and associated with \code{environment}, \code{user.id}, and \code{user.pass}. Currently supported repositories are: "EDI" (Environmental Data Initiative). Requests for support of other repositories can be made via \href{https://github.com/EDIorg/ecocomDP}{ecocomDP GitHub} issues. Default is "EDI".}

\item{evironment}{(character) Repository environment in which the L0 and L1 exist. Some repositories have development, staging, and production environments which are distinct from one another. This argument allows execution for the \code{update_L1} workflow within the context of one of these environments. Default is "production".}

\item{path}{(character) Directory to which L1 tables, scripts, and metadata will be written.}

\item{url}{(character) Publicly accessible URL to \code{path} for download by a data repository.}

\item{user.id}{(character) User identifier within a specified \code{repository}. This controls editing access in some \code{repository}.}

\item{user.pass}{(character) Password associated with \code{user.id} for repository upload.}
}
\description{
Updates an L1 data package when it’s L0 parent data package has been updated. This function is a wrapper to several subroutines.
}
\details{
The updated L1 data package published in the data repository system specified by \code{repository} and in the sub-system specified by \code{environment}. Messaging from this routine is written to log_master.txt. Dependencies specified in conversion scripts are added to dependencies_update_L1.txt and installed in servers R distribution via \code{update_supported_libraries()}.

\code{user.id}, \code{repository}, and \code{user.pass} form a set of credentials allowing edits to all data packages within a repository system, otherwise evaluate/upload will fail if the creator is someone else.

Requires an L0 and L1 already exist in the repository, because the L1 creation and upload process may have errors that need resolution by a human.
}
