% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_column_names.R
\name{validate_column_names}
\alias{validate_column_names}
\title{validate_column_names}
\usage{
validate_column_names(data.path, criteria)
}
\arguments{
\item{tables}{(character) Names of valid L1 tables in your dataset working directory.
Get valid table names with `validate_table_names`.}

\item{data.path}{(character) Path to the directory containing the L1 tables.}

\item{criteria}{(data frame) Validation criteria located at 
/inst/validation_criteria.txt.}
}
\value{
If required column names are missing, then an error is returned.
}
\description{
Check that column names of L1 tables are valid.
}
