% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_column_names}
\alias{validate_column_names}
\title{Check for invalid column names}
\usage{
validate_column_names(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If column names outside of the allowed set are found, then 
    invalid column names are returned, otherwise NULL is returned.
}
\description{
Check for invalid column names
}
