% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_column_presence}
\alias{validate_column_presence}
\title{Check for required columns}
\usage{
validate_column_presence(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If required column names are absent, then missing columns
    are returned, otherwise NULL is returned.
}
\description{
Check for required columns
}
