% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_column_presence.R
\name{validate_column_presence}
\alias{validate_column_presence}
\title{Validate column presence}
\usage{
validate_column_presence(data.path)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}
}
\value{
A validation report printed in the console window of RStudio. The 
    validation checks run until an error is encountered. You must address 
    errors before continuing to the next check. Once all validation checks 
    have successfully completed, you will be notified with a congratulatory 
    message.
}
\description{
This function ensures that your ecocomDP (L1) tables contain 
    required columns.
}
\details{
The full suite of L1 validation checks are performed by the 
   \code{validate_ecocomDP} function. The sequence of checks performed in
   \code{validate_ecocomDO} are not random, rather some checks are dependent
   upon others.
}
