% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_composite_keys.R
\name{validate_composite_keys}
\alias{validate_composite_keys}
\title{validate_composite_keys}
\usage{
validate_composite_keys(tables, data.path, criteria)
}
\arguments{
\item{tables}{(character) Names of valid L1 tables in your dataset working directory.
Get valid table names with `validate_table_names`.}

\item{data.path}{(character) Path to the directory containing the L1 tables.}

\item{criteria}{(data frame) Validation criteria located at 
/inst/validation_criteria.txt.}
}
\value{
If composite keys are not unique, then an error is returned.
}
\description{
Check that composite keys of a table are unique and issue an error if
    otherwise.
}
