% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_datetime.R
\name{validate_datetime}
\alias{validate_datetime}
\title{validate_datetime}
\usage{
validate_datetime(tables, data.path, criteria)
}
\arguments{
\item{tables}{(character) Names of valid L1 tables in your dataset working directory.
Get valid table names with `validate_table_names`.}

\item{data.path}{(character) Path to the directory containing the L1 tables.}

\item{criteria}{(data frame) Validation criteria located at 
/inst/validation_criteria.txt.}
}
\value{
Return an error if datetime formats are invalid.
}
\description{
Validate datetime formats of input file. Valid datetime format is
    YYYY-MM-DD hh:mm:ss
}
