% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_ecocomDP}
\alias{validate_ecocomDP}
\title{validate_ecocomDP}
\usage{
validate_ecocomDP(data.path = NULL, data.list = NULL)
}
\arguments{
\item{data.path}{(character) The path to the directory containing L1
tables.}

\item{data.list}{(list of data frames) A named list of data frames containing the L1
tables. Data frame tables must be named after the file name from 
which they were read, including the extension.}
}
\value{
A validation report to the RStudio console window. When an issue is 
    encountered a message will be displayed with recommendations for fixing
    it. Each issue must be resolved before the next check will be implemented.
    Once all validation checks have been passed, you will be notified with 
    a congratulatory message and you can move on to creating EML metadata with
    \code{make_eml}.
}
\description{
Once you've created an ecocomDP (L1) you will need to check that it 
    is accurately formatted. This validation process ensures your L1 is 
    correct and can be combined with other L1.
}
\details{
Run this function after creating a L1 and before making EML for it.

   Validation checks performed by this function:
   \itemize{
       \item \strong{Table names} Table names must follow the ecocomDP 
       naming convention (i.e. \emph{studyName_ecocomDPTableName.ext}, e.g. 
       \emph{gleon_chloride_observation.csv}). 
       \item \strong{Table presence} Some L1 tables are 
       required, others are not.
       \item \strong{Column names} Column names must be those specified by
       the data pattern.
       \item \strong{Column presence} Some columns are 
       required, others are not.
       \item \strong{Datetime format} One datetime format is valid, namely
       'YYYY-MM-DDThh:mm:ss'.
       \item \strong{Column classes} Column classes must match the 
       ecocomDP specification.
       \item \strong{Primary keys} Primary keys must be valid.
       \item \strong{Composite keys} Composite keys must be valid.
       \item \strong{Referential integrity} Referential integrity must be
       valid.
   }
}
