% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R, R/validate_table_ids.R
\name{validate_ecocomDP}
\alias{validate_ecocomDP}
\alias{validate_ecocomDP}
\title{Validate ecocomDP tables}
\usage{
validate_ecocomDP(data.path, sep)

validate_table_ids(data.path)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}

\item{data.path}{A character string specifying the path to the directory containing L1
tables.}
}
\value{
A validation report to the RStudio console window. When an issue is 
    encountered a message will be displayed with recommendations for fixing
    it. Each issue must be resolved before the next check will be implemented.
    Once all validation checks have been passed, you will be notified with 
    a congratulatory message and you can move on to creating EML metadata with
    \code{make_eml}.

A validation report printed in the console window of RStudio. The 
    validation checks run until an error is encountered. You must address 
    errors before continuing to the next check. Once all validation checks 
    have successfully completed, you will be notified with a congratulatory 
    message.
}
\description{
Once you've created an ecocomDP (L1) you will need to check that it 
    is accurately formatted. This validation process ensures your L1 is 
    correct and can be combined with other L1.

This function ensures that your ecocomDP (L1) tables contain valid 
    relational IDs (keys).
}
\details{
Run this function after creating a L1 and before making EML for it.

   Validation checks performed by this function:
   \itemize{
       \item \strong{Table names} Table names must follow the ecocomDP 
       naming convention (i.e. \emph{studyName_ecocomDPTableName.ext}, e.g. 
       \emph{gleon_chloride_observation.csv}). 
       \item \strong{Required tables} Some L1 tables are 
       required, others are not.
       \item \strong{Column names} Column names must be those specified by
       the data pattern.
       \item \strong{Required columns} Some columns are 
       required, others are not.
       \item \strong{Column classes} Column classes must match the 
       ecocomDP specification.
       \item \strong{Datetime format string} Datetime values should match
       the input datetime.format.string parameter.
   }

The full suite of L1 validation checks are performed by the 
   \code{validate_ecocomDP} function. The sequence of checks performed in
   \code{validate_ecocomDO} are not random, rather some checks are dependent
   upon others.
}
