% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_ecocomDP}
\alias{validate_ecocomDP}
\title{Validate a dataset against the ecocomDP}
\usage{
validate_ecocomDP(data.path = NULL, data.list = NULL)
}
\arguments{
\item{data.path}{(character) The path to the directory containing ecocomDP tables.}

\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table. See output from \code{read_data()}.}
}
\description{
Use this function to verify a dataset conforms to the ecocomDP.
}
\details{
Validation checks:
   \itemize{
       \item{File names - When using \code{data.path} the data are read from
       files and must follow an expected pattern, namely 
       \emph{studyName_ecocomDPTableName.ext} (e.g. 
       \emph{gleon_chloride_observation.csv}).}
       \item{Table presence - Some tables are required, others are not.}
       \item{Column names - Column names must follow specification.}
       \item{Column presence - Some columns are required, others are not.}
       \item{Column classes - Column classes must follow specification.}
       \item{Datetime format - Date and datetime formats must follow 
       specification.}
       \item{Primary keys - Primary keys of tables must be valid.}
       \item{Composite keys - Composite keys of tables must be valid.}
       \item{Referential integrity - Referential integrity among tables must 
       be valid.}
   }
}
\note{
This function is used by ecocomDP creators (to ensure what has been 
    created is valid), maintainers (to improve the quality of archived
    ecocomDP datasets), and users (to ensure the data being used is free of
    model error).
}
\examples{
# Validate a set files (example data)
validate_ecocomDP(data.path = system.file("/data", package = "ecocomDP"))
        
}
