% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_ecocomDP}
\alias{validate_ecocomDP}
\title{Validate ecocomDP tables}
\usage{
validate_ecocomDP(data.path, sep)
    Run this function after creating an ecocomDP and before making EML for 
    it.
}
\arguments{
\item{data.path}{A path to the dataset working directory containing \emph{only} ecocomDP 
tables. This directory should not contain anything else.}

\item{sep}{The field delimiter of ecocomDP tables. Can be comma or tab delimited 
(i.e. "," or "\\t")}
}
\value{
A validation report printed in the console window of RStudio. This 
    validator runs until it hits an error. You must address the reported
    issues before continuing. Once all the validation checks have been 
    passed, you will recieve a congradulatory message.
}
\description{
Once you've formatted a dataset to ecocomDP you will need to check that 
    it complies with the data pattern standard. This process ensures your 
    dataset is valid and can be combined with other datasets in the ecocomDP
    with the aggregate_ecocomDP function.
}
\details{
Checks:
   \itemize{
       \item \strong{Table naming} Table names must follow the ecocomDP 
       naming convention (i.e. \emph{studyName_ecocomDPTableName}). Make 
       sure table names are consistent.
       \item \strong{Required tables} Some tables are required. Add any 
       missing tables.
       \item \strong{Column names} Column names must be valid.
       Adjust column names as required.
       \item \strong{Required columns} Some columns are required.
       Add missing columns.
       \item \strong{Column classes} Column classes must be correct.
   }
}
