% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file_names.R
\name{validate_file_names}
\alias{validate_file_names}
\title{Validate file names}
\usage{
validate_file_names(path = "", data.files = data.files = c("data.file.1", 
    "data.file.2", "etc.")).
}
\arguments{
\item{path}{A character string specifying a path to the dataset working directory.}

\item{data.files}{A list of character strings specifying the names of the data files
of your dataset. It is not necessary to include the file extension.}
}
\value{
A warning message if the data files don't exist at path, and which of
    the input data files are missing.
    
    The full names of files listed in the data.files argument.
}
\description{
Identify whether input data file names exist in the specified directory.
}
