% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_primary_keys}
\alias{validate_primary_keys}
\title{Check primary keys}
\usage{
validate_primary_keys(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If primary keys are not unique, then row numbers of 
    duplicate keys are returned, otherwise NULL is returned.
}
\description{
Check primary keys
}
