% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_referential_integrity}
\alias{validate_referential_integrity}
\title{Check referential integrity}
\usage{
validate_referential_integrity(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If foreign keys do not match primary keys, then unmatched 
    foreign keys are returned, otherwise NULL is returned.
}
\description{
Check referential integrity
}
