% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_referential_integrity.R
\name{validate_referential_integrity}
\alias{validate_referential_integrity}
\title{validate_referential_integrity}
\usage{
validate_referential_integrity(tables, data.path, criteria)
}
\arguments{
\item{tables}{(character) Names of valid L1 tables in your dataset working directory.
Get valid table names with `validate_table_names`.}

\item{data.path}{(character) Path to the directory containing the L1 tables.}

\item{criteria}{(data frame) Validation criteria located at 
/inst/validation_criteria.txt.}
}
\value{
If foreign keys do not match primary keys, then an error is returned.
}
\description{
Check that foreign keys match primary keys.
}
