% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_validation_criteria.R
\name{validate_table_names}
\alias{validate_table_names}
\title{Load validation criteria}
\usage{
load_validation_criteria(data.path)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}
}
\description{
Load criteria against which user supplied ecocomDP tables are validated.
}
\details{
This function loads and parses validation criteria from 
   \emph{validation_criteria.txt} located in the \emph{\inst} directory of
   the \code{ecocomDP} package.
}
