% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_validation_criteria.R,
%   R/validate_table_names.R
\name{validate_table_names}
\alias{validate_table_names}
\alias{validate_table_names}
\title{Load validation criteria}
\usage{
load_validation_criteria(data.path)

validate_table_names(data.path)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}

\item{data.path}{A character string specifying the path to the directory containing L1
tables.}
}
\value{
A validation report printed in the RStudio console window. The 
    validation check runs until an error is encountered. You must address 
    errors of one table before the next table is checked. Once all 
    tables have been checked you will be notified with a congratulatory 
    message.
}
\description{
Load criteria against which user supplied ecocomDP tables are validated.

This function ensures that your ecocomDP (L1) tables follow the
    table naming convention (i.e. \emph{studyName_ecocomDPTableName.ext}, 
    e.g. \emph{gleon_chloride_observation.csv}).
}
\details{
This function loads and parses validation criteria from 
   \emph{validation_criteria.txt} located in the \emph{\inst} directory of
   the \code{ecocomDP} package.

The full suite of L1 validation checks are performed by the 
   \code{validate_ecocomDP} function. The sequence of checks performed in
   \code{validate_ecocomDO} are not random, rather some checks are dependent
   upon others.
}
