% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_validation_criteria.R,
%   R/validate_ecocomDP.R
\name{validate_table_names}
\alias{validate_table_names}
\alias{validate_table_names}
\title{Load validation criteria}
\usage{
load_validation_criteria(data.path)

validate_table_names(data.path, criteria, data.list = NULL)
}
\arguments{
\item{data.path}{A character string specifying the path to the directory containing L1
tables.}

\item{criteria}{(data frame) Validation criteria located in the R package inst directory.}

\item{data.list}{(list of data frames) A list of of data frames with the following 
structure:
\itemize{
    \item \strong{list name}
    \itemize{
        \item \strong{ecocomDP table name} i.e. 
    }
    \item \strong{L1 table name:} i.e. observation, location, taxon, etc.
        \item \strong{data frame:} Name data frame Data frame representation of the L1 table
        \item \strong{file name} File name the data frame was created from.
}}

\item{data.path}{(character) Path to the directory containing the L1 tables.
tables. Use this argument if tables are locally stored and need to be
read in to R.}
}
\value{
If table names are valid, then the corresponding table names are
    returned. If table names are invalid, then an error message is
    returned.
}
\description{
Load criteria against which user supplied ecocomDP tables are validated.

This function ensures that your ecocomDP (L1) tables follow the
    table naming convention (i.e. \emph{studyName_ecocomDPTableName.ext},
    e.g. \emph{gleon_chloride_observation.csv}).
}
\details{
This function loads and parses validation criteria from 
   \emph{validation_criteria.txt} located in the \emph{\inst} directory of
   the \code{ecocomDP} package.
}
