% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ecocomDP.R
\name{validate_table_names}
\alias{validate_table_names}
\title{Check for file name errors}
\usage{
validate_table_names(data.path = NULL)
}
\arguments{
\item{data.path}{(character) The path to the directory containing ecocomDP tables.}
}
\value{
If table names are valid, then the corresponding table names are
    returned. If table names are invalid, then an error message is
    returned.
}
\description{
This function ensures that your ecocomDP (L1) tables follow the
    file naming convention (i.e. \emph{studyName_ecocomDPTableName.ext},
    e.g. \emph{gleon_chloride_observation.csv}).
}
