% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_presence.R
\name{validate_table_presence}
\alias{validate_table_presence}
\title{validate_table_presence}
\usage{
validate_table_presence(tables, criteria)
}
\arguments{
\item{tables}{(character) Names of valid L1 tables in your dataset working directory.
Get valid table names with `validate_table_names`.}

\item{criteria}{(data frame) Validation criteria located at 
/inst/validation_criteria.txt.}
}
\value{
If required tables are missing, then an error message is produced.
}
\description{
Check that required L1 tables are present.
}
\details{

}
