% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_table_presence}
\alias{validate_table_presence}
\title{Check for required tables}
\usage{
validate_table_presence(data.list)
}
\arguments{
\item{data.list}{(list of data frames) A named list of data frames, each of which is an 
ecocomDP table.}
}
\value{
(character) If required tables are missing, then the missing table names
    are returned, otherwise NULL is returned.
}
\description{
Check for required tables
}
