% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{view_descriptions}
\alias{view_descriptions}
\title{View descriptions and requirements of ecocomDP tables}
\usage{
view_descriptions()
}
\value{
(NULL) Opens a webpage, in your default browser, with a list of descriptions and requirements of the ecocomDP tables
}
\description{
View descriptions and requirements of ecocomDP tables
}
\examples{
\dontrun{
view_descriptions()
}

}
