% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{view_diagram}
\alias{view_diagram}
\title{View diagram of ecocomDP tables and relationships}
\usage{
view_diagram()
}
\value{
(NULL) Opens a webpage, in your default browser, with a diagram keys and linkages among ecocomDP tables
}
\description{
View diagram of ecocomDP tables and relationships
}
\examples{
\dontrun{
view_diagram()
}

}
