% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_landing_page.R
\name{view_landing_page}
\alias{view_landing_page}
\title{view_landing_page}
\usage{
view_landing_page(package.id)
}
\arguments{
\item{package.id}{(character) The data package ID listed in the data frame created by 
`list_all_ecocomDP`. EDI data package IDs are of the form 
'edi.package_number.revision_number' (e.g. edi.100.2). NEON data package
IDs are of the form 'data_product.identifier.revision_number' (e.g.
'DP1.20120.001').}
}
\value{
The data package landing page.
}
\description{
View the human readable landing page of a data package EML record.
}
