% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{view_model_diagram}
\alias{view_model_diagram}
\title{View ecocomDP model diagram}
\usage{
view_model_diagram()
}
\description{
View ecocomDP model diagram
}
\examples{
\dontrun{
view_model_diagram()
}

}
